remove couple more files
