+++ docs/assets/js/archive.js
@@ -1,1 +1,1 @@
-function t(t,e){return"string"==typeof t?(e||document).querySelector(t):t||null}function e(t){let e=t.getBoundingClientRect();return{top:e.top+(document.documentElement.scrollTop||document.body.scrollTop),left:e.left+(document.documentElement.scrollLeft||document.body.scrollLeft)}}t.create=(e,s)=>{var i=document.createElement(e);for(var a in s){var n=s[a];if("inside"===a)t(n).appendChild(i);else if("around"===a){var r=t(n);r.parentNode.insertBefore(i,r),i.appendChild(r)}else"styles"===a?"object"==typeof n&&Object.keys(n).map((t=>{i.style[t]=n[t]})):a in i?i[a]=n:i.setAttribute(a,n)}return i};const s={margins:{top:10,bottom:10,left:20,right:20},paddings:{top:20,bottom:40,left:30,right:10},baseHeight:240,titleHeight:20,legendHeight:30,titleFontSize:12};function i(t){return t.titleHeight+t.margins.top+t.paddings.top}function a(t){return t.margins.left+t.paddings.left}function n(t){return t.margins.top+t.margins.bottom+t.paddings.top+t.paddings.bottom+t.titleHeight+t.legendHeight}function r(t){return t.margins.left+t.margins.right+t.paddings.left+t.paddings.right}const o=["light-blue","blue","violet","red","orange","yellow","green","light-green","purple","magenta","light-grey","dark-grey"],l={bar:o,line:o,pie:o,percentage:o,heatmap:["#ebedf0","#c6e48b","#7bc96f","#239a3b","#196127"],donut:o},h=Math.PI/180;class d{constructor({parent:t=null,colors:e=[]}){this.parent=t,this.colors=e,this.titleName="",this.titleValue="",this.listValues=[],this.titleValueFirst=0,this.x=0,this.y=0,this.top=0,this.left=0,this.setup()}setup(){this.makeTooltip()}refresh(){this.fill(),this.calcPosition()}makeTooltip(){this.container=t.create("div",{inside:this.parent,className:"graph-svg-tip comparison",innerHTML:'<span class="title"></span>\n\t\t\t\t<ul class="data-point-list"></ul>\n\t\t\t\t<div class="svg-pointer"></div>'}),this.hideTip(),this.title=this.container.querySelector(".title"),this.dataPointList=this.container.querySelector(".data-point-list"),this.parent.addEventListener("mouseleave",(()=>{this.hideTip()}))}fill(){let e;this.index&&this.container.setAttribute("data-point-index",this.index),e=this.titleValueFirst?`<strong>${this.titleValue}</strong>${this.titleName}`:`${this.titleName}<strong>${this.titleValue}</strong>`,this.title.innerHTML=e,this.dataPointList.innerHTML="",this.listValues.map(((e,s)=>{const i=this.colors[s]||"black";let a=0===e.formatted||e.formatted?e.formatted:e.value,n=t.create("li",{styles:{"border-top":`3px solid ${i}`},innerHTML:`<strong style="display: block;">${0===a||a?a:""}</strong>\n\t\t\t\t\t${e.title?e.title:""}`});this.dataPointList.appendChild(n)}))}calcPosition(){let t=this.container.offsetWidth;this.top=this.y-this.container.offsetHeight-5,this.left=this.x-t/2;let e=this.parent.offsetWidth-t,s=this.container.querySelector(".svg-pointer");if(this.left<0)s.style.left=`calc(50% - ${-1*this.left}px)`,this.left=0;else if(this.left>e){let t=`calc(50% + ${this.left-e}px)`;s.style.left=t,this.left=e}else s.style.left="50%"}setValues(t,e,s={},i=[],a=-1){this.titleName=s.name,this.titleValue=s.value,this.listValues=i,this.x=t,this.y=e,this.titleValueFirst=s.valueFirst||0,this.index=a,this.refresh()}hideTip(){this.container.style.top="0px",this.container.style.left="0px",this.container.style.opacity="0"}showTip(){this.container.style.top=this.top+"px",this.container.style.left=this.left+"px",this.container.style.opacity="1"}}function c(t){return parseFloat(t.toFixed(2))}function p(t,e,s,i=!1){s||(s=i?t[0]:t[t.length-1]);let a=new Array(Math.abs(e)).fill(s);return t=i?a.concat(t):t.concat(a)}function u(t,e){return(t+"").length*e}function m(t,e){return{x:Math.sin(t*h)*e,y:Math.cos(t*h)*e}}function g(t,e=!1){return!Number.isNaN(t)&&(void 0!==t&&(!!Number.isFinite(t)&&!(e&&t<0)))}function f(t){let e,s,i;if(t instanceof Date)return new Date(t.getTime());if("object"!=typeof t||null===t)return t;for(i in e=Array.isArray(t)?[]:{},t)s=t[i],e[i]=f(s);return e}function y(t,e){let s,i;return t<=e?(s=e-t,i=t):(s=t-e,i=e),[s,i]}function b(t,e,s=e.length-t.length){return s>0?t=p(t,s):e=p(e,s),[t,e]}function x(t,e){if(t)return t.length>e?t.slice(0,e-3)+"...":t}function v(t,e){let s=[];for(let i=0;i<t.length;i++)s.push([t[i],e[i]]);let i=(t,e,s,i)=>{let a=((t,e)=>{let s=e[0]-t[0],i=e[1]-t[1];return{length:Math.sqrt(Math.pow(s,2)+Math.pow(i,2)),angle:Math.atan2(i,s)}})(e||t,s||t),n=a.angle+(i?Math.PI:0),r=.2*a.length;return[t[0]+Math.cos(n)*r,t[1]+Math.sin(n)*r]};return((t,e)=>t.reduce(((t,s,i,a)=>0===i?`${s[0]},${s[1]}`:`${t} ${e(s,i,a)}`),""))(s,((t,e,s)=>{let a=i(s[e-1],s[e-2],t),n=i(t,s[e-1],s[e+1],!0);return`C ${a[0]},${a[1]} ${n[0]},${n[1]} ${t[0]},${t[1]}`}))}const w={"light-blue":"#7cd6fd",blue:"#5e64ff",violet:"#743ee2",red:"#ff5858",orange:"#ffa00a",yellow:"#feef72",green:"#28a745","light-green":"#98d85b",purple:"#b554ff",magenta:"#ffa3ef",black:"#36114C",grey:"#bdd3e6","light-grey":"#f0f4f7","dark-grey":"#b8c2cc"};function A(t){return t>255?255:t<0?0:t}function k(t,e){let s=L(t),i=!1;"#"==s[0]&&(s=s.slice(1),i=!0);let a=parseInt(s,16),n=A((a>>16)+e),r=A((a>>8&255)+e);return(i?"#":"")+(A((255&a)+e)|r<<8|n<<16).toString(16)}const L=t=>/rgb[a]{0,1}\([\d, ]+\)/gim.test(t)?/\D+(\d*)\D+(\d*)\D+(\d*)/gim.exec(t).map(((t,e)=>0!==e?Number(t).toString(16):"#")).reduce(((t,e)=>`${t}${e}`)):w[t]||t;function T(t,e){return"string"==typeof t?(e||document).querySelector(t):t||null}function M(t,e){var s=document.createElementNS("http://www.w3.org/2000/svg",t);for(var i in e){var a=e[i];if("inside"===i)T(a).appendChild(s);else if("around"===i){var n=T(a);n.parentNode.insertBefore(s,n),s.appendChild(n)}else"styles"===i?"object"==typeof a&&Object.keys(a).map((t=>{s.style[t]=a[t]})):("className"===i&&(i="class"),"innerHTML"===i?s.textContent=a:s.setAttribute(i,a))}return s}function D(t,e,s,i){return M("stop",{inside:t,style:`stop-color: ${s}`,offset:e,"stop-opacity":i})}function P(t,e="",s){let i={className:t,transform:e};return s&&(i.inside=s),M("g",i)}function C(t,e="",s="none",i="none",a=2){return M("path",{className:e,d:t,styles:{stroke:s,fill:i,"stroke-width":a}})}function $(t,e,s=!1){let i="path-fill-gradient-"+e+"-"+(s?"lighter":"default"),a=function(t,e){return M("linearGradient",{inside:t,id:e,x1:0,x2:0,y1:0,y2:1})}(t,i),n=[1,.6,.2];return s&&(n=[.4,.2,0]),D(a,"0%",e,n[0]),D(a,"50%",e,n[1]),D(a,"100%",e,n[2]),i}function N(t,e,s,i,a,n="none",r={}){let o={className:t,x:e,y:s,width:i,height:i,rx:a,fill:n};return Object.keys(r).map((t=>{o[t]=r[t]})),M("rect",o)}function E(t,e,s,i,a={}){let n=a.fontSize||10;return M("text",{className:t,x:e,y:s,dy:(void 0!==a.dy?a.dy:n/2)+"px","font-size":n+"px",fill:a.fill||"#555b51","text-anchor":a.textAnchor||"start",innerHTML:i})}function O(t,e,s,i,a={}){a.stroke||(a.stroke="#dadada"),a.lineType||(a.lineType=""),a.shortenNumbers&&(e=function(t){let e;if("number"==typeof t)e=t;else if("string"==typeof t&&(e=Number(t),Number.isNaN(e)))return t;let s=Math.floor(Math.log10(Math.abs(e)));if(s<=2)return e;let i=Math.floor(s/3),a=Math.pow(10,s-3*i)*+(e/Math.pow(10,s)).toFixed(1);return Math.round(100*a)/100+" "+["","K","M","B","T"][i]}(e));let n=M("line",{className:"line-horizontal "+a.className+("dashed"===a.lineType?"dashed":""),x1:s,x2:i,y1:0,y2:0,styles:{stroke:a.stroke}}),r=M("text",{x:s<i?s-4:s+4,y:0,dy:"3px","font-size":"10px","text-anchor":s<i?"end":"start",innerHTML:e+""}),o=M("g",{transform:`translate(0, ${t})`,"stroke-opacity":1});return 0!==r&&"0"!==r||(o.style.stroke="rgba(27, 31, 35, 0.6)"),o.appendChild(n),o.appendChild(r),o}function S(t,e,s,i={}){g(t)||(t=0),i.pos||(i.pos="bottom"),i.offset||(i.offset=0),i.mode||(i.mode="span"),i.stroke||(i.stroke="#dadada"),i.className||(i.className="");let a=s+6,n="span"===i.mode?-6:s;return"tick"===i.mode&&"top"===i.pos&&(a=-6,n=0),function(t,e,s,i,a={}){a.stroke||(a.stroke="#dadada");let n=M("line",{className:"line-vertical "+a.className,x1:0,x2:0,y1:s,y2:i,styles:{stroke:a.stroke}}),r=M("text",{x:0,y:s>i?s+4:s-4-10,dy:"10px","font-size":"10px","text-anchor":"middle",innerHTML:e+""}),o=M("g",{transform:`translate(${t}, 0)`});return o.appendChild(n),o.appendChild(r),o}(t,e,a,n,{stroke:i.stroke,className:i.className,lineType:i.lineType})}let z={bar:t=>{let e;"rect"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode();return s.style.fill="#000000",s.style.opacity="0.4",e&&s.setAttribute("transform",e),s},dot:t=>{let e;"circle"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode(),i=t.getAttribute("r"),a=t.getAttribute("fill");return s.setAttribute("r",parseInt(i)+4),s.setAttribute("fill",a),s.style.opacity="0.6",e&&s.setAttribute("transform",e),s},heat_square:t=>{let e;"circle"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode(),i=t.getAttribute("r"),a=t.getAttribute("fill");return s.setAttribute("r",parseInt(i)+4),s.setAttribute("fill",a),s.style.opacity="0.6",e&&s.setAttribute("transform",e),s}},W={bar:(t,e)=>{let s;"rect"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["x","y","width","height"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)},dot:(t,e)=>{let s;"circle"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["cx","cy"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)},heat_square:(t,e)=>{let s;"circle"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["cx","cy"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)}};function H(t,e,s,i){let a="string"==typeof e?e:e.join(", ");return[t,{transform:s.join(", ")},i,"easein","translate",{transform:a}]}function F(t,e,s){return H(t,[0,s],[0,e],350)}function R(t,e){return[t,{d:e},350,"easein"]}const I={ease:"0.25 0.1 0.25 1",linear:"0 0 1 1",easein:"0.1 0.8 0.2 1",easeout:"0 0 0.58 1",easeinout:"0.42 0 0.58 1"};function Y(t,e){t.style.transform=e,t.style.webkitTransform=e,t.style.msTransform=e,t.style.mozTransform=e,t.style.oTransform=e}function j(t,e){let s=[],i=[];e.map((t=>{let e,a,n=t[0],r=n.parentNode;t[0]=n,[e,a]=function(t,e,s,i="linear",a,n={}){let r=t.cloneNode(!0),o=t.cloneNode(!0);for(var l in e){let d;d="transform"===l?document.createElementNS("http://www.w3.org/2000/svg","animateTransform"):document.createElementNS("http://www.w3.org/2000/svg","animate");let c=n[l]||t.getAttribute(l),p=e[l],u={attributeName:l,from:c,to:p,begin:"0s",dur:s/1e3+"s",values:c+";"+p,keySplines:I[i],keyTimes:"0;1",calcMode:"spline",fill:"freeze"};for(var h in a&&(u.type=a),u)d.setAttribute(h,u[h]);r.appendChild(d),a?o.setAttribute(l,`translate(${p})`):o.setAttribute(l,p)}return[r,o]}(...t),s.push(a),i.push([e,r]),r&&r.replaceChild(e,n)}));let a=t.cloneNode(!0);return i.map(((t,i)=>{t[1]&&(t[1].replaceChild(s[i],t[0]),e[i][0]=s[i])})),a}class B{constructor(t,e){if(e=f(e),this.parent="string"==typeof t?document.querySelector(t):t,!(this.parent instanceof HTMLElement))throw new Error("No `parent` element to render on was provided.");this.rawChartArgs=e,this.title=e.title||"",this.type=e.type||"",this.realData=this.prepareData(e.data),this.data=this.prepareFirstData(this.realData),this.colors=this.validateColors(e.colors,this.type),this.config={showTooltip:1,showLegend:1,isNavigable:e.isNavigable||0,animate:void 0!==e.animate?e.animate:1,truncateLegends:e.truncateLegends||1},this.measures=JSON.parse(JSON.stringify(s));let i=this.measures;this.setMeasures(e),this.title.length||(i.titleHeight=0),this.config.showLegend||(i.legendHeight=0),this.argHeight=e.height||i.baseHeight,this.state={},this.options={},this.initTimeout=700,this.config.isNavigable&&(this.overlays=[]),this.configure(e)}prepareData(t){return t}prepareFirstData(t){return t}validateColors(t,e){const s=[];return(t=(t||[]).concat(l[e])).forEach((t=>{const e=L(t);!function(t){return/(^\s*)(#)((?:[A-Fa-f0-9]{3}){1,2})$/i.test(t)||/(^\s*)(rgb|hsl)(a?)[(]\s*([\d.]+\s*%?)\s*,\s*([\d.]+\s*%?)\s*,\s*([\d.]+\s*%?)\s*(?:,\s*([\d.]+)\s*)?[)]$/i.test(t)}(e)?console.warn('"'+t+'" is not a valid color.'):s.push(e)})),s}setMeasures(){}configure(){let t=this.argHeight;this.baseHeight=t,this.height=t-n(this.measures),this.boundDrawFn=()=>this.draw(!0),ResizeObserver&&(this.resizeObserver=new ResizeObserver(this.boundDrawFn),this.resizeObserver.observe(this.parent)),window.addEventListener("resize",this.boundDrawFn),window.addEventListener("orientationchange",this.boundDrawFn)}destroy(){this.resizeObserver&&this.resizeObserver.disconnect(),window.removeEventListener("resize",this.boundDrawFn),window.removeEventListener("orientationchange",this.boundDrawFn)}setup(){this.makeContainer(),this.updateWidth(),this.makeTooltip(),this.draw(!1,!0)}makeContainer(){this.parent.innerHTML="";let e={inside:this.parent,className:"chart-container"};this.independentWidth&&(e.styles={width:this.independentWidth+"px"}),this.container=t.create("div",e)}makeTooltip(){this.tip=new d({parent:this.container,colors:this.colors}),this.bindTooltip()}bindTooltip(){}draw(t=!1,e=!1){t&&null===this.parent.offsetParent||(this.updateWidth(),this.calc(t),this.makeChartArea(),this.setupComponents(),this.components.forEach((t=>t.setup(this.drawArea))),this.render(this.components,!1),e&&(this.data=this.realData,setTimeout((()=>{this.update(this.data)}),this.initTimeout)),this.renderLegend(),this.setupNavigation(e))}calc(){}updateWidth(){var t,e,s;this.baseWidth=(t=this.parent,e=window.getComputedStyle(t),s=parseFloat(e.paddingLeft)+parseFloat(e.paddingRight),t.clientWidth-s),this.width=this.baseWidth-r(this.measures)}makeChartArea(){this.svg&&this.container.removeChild(this.svg);let t=this.measures;var e,s,n,r;this.svg=(e=this.container,s="frappe-chart chart",n=this.baseWidth,r=this.baseHeight,M("svg",{className:s,inside:e,width:n,height:r})),this.svgDefs=M("defs",{inside:this.svg}),this.title.length&&(this.titleEL=E("title",t.margins.left,t.margins.top,this.title,{fontSize:t.titleFontSize,fill:"#666666",dy:t.titleFontSize}));let o=i(t);this.drawArea=P(this.type+"-chart chart-draw-area",`translate(${a(t)}, ${o})`),this.config.showLegend&&(o+=this.height+t.paddings.bottom,this.legendArea=P("chart-legend",`translate(${a(t)}, ${o})`)),this.title.length&&this.svg.appendChild(this.titleEL),this.svg.appendChild(this.drawArea),this.config.showLegend&&this.svg.appendChild(this.legendArea),this.updateTipOffset(a(t),i(t))}updateTipOffset(t,e){this.tip.offset={x:t,y:e}}setupComponents(){this.components=new Map}update(t){t||console.error("No data to update."),this.data=this.prepareData(t),this.calc(),this.render(this.components,this.config.animate),this.renderLegend()}render(t=this.components,e=!0){this.config.isNavigable&&this.overlays.map((t=>t.parentNode.removeChild(t)));let s=[];t.forEach((t=>{s=s.concat(t.update(e))})),s.length>0?(!function(t,e,s){if(0===s.length)return;let i=j(e,s);e.parentNode==t&&(t.removeChild(e),t.appendChild(i)),setTimeout((()=>{i.parentNode==t&&(t.removeChild(i),t.appendChild(e))}),250)}(this.container,this.svg,s),setTimeout((()=>{t.forEach((t=>t.make())),this.updateNav()}),400)):(t.forEach((t=>t.make())),this.updateNav())}updateNav(){this.config.isNavigable&&(this.makeOverlay(),this.bindUnits())}renderLegend(){}setupNavigation(t=!1){this.config.isNavigable&&t&&(this.bindOverlay(),this.keyActions={13:this.onEnterKey.bind(this),37:this.onLeftArrow.bind(this),38:this.onUpArrow.bind(this),39:this.onRightArrow.bind(this),40:this.onDownArrow.bind(this)},document.addEventListener("keydown",(t=>{var e,s;e=this.container,(s=e.getBoundingClientRect()).top>=0&&s.left>=0&&s.bottom<=(window.innerHeight||document.documentElement.clientHeight)&&s.right<=(window.innerWidth||document.documentElement.clientWidth)&&(t=t||window.event,this.keyActions[t.keyCode]&&this.keyActions[t.keyCode]())})))}makeOverlay(){}updateOverlay(){}bindOverlay(){}bindUnits(){}onLeftArrow(){}onRightArrow(){}onUpArrow(){}onDownArrow(){}onEnterKey(){}addDataPoint(){}removeDataPoint(){}getDataPoint(){}setCurrentDataPoint(){}updateDataset(){}export(){let e=function(e){let s=e.cloneNode(!0);s.classList.add("chart-container"),s.setAttribute("xmlns","http://www.w3.org/2000/svg"),s.setAttribute("xmlns:xlink","http://www.w3.org/1999/xlink");let i=t.create("style",{innerHTML:".chart-container{position:relative;font-family:-apple-system,BlinkMacSystemFont,'Segoe UI','Roboto','Oxygen','Ubuntu','Cantarell','Fira Sans','Droid Sans','Helvetica Neue',sans-serif}.chart-container .axis,.chart-container .chart-label{fill:#555b51}.chart-container .axis line,.chart-container .chart-label line{stroke:#dadada}.chart-container .dataset-units circle{stroke:#fff;stroke-width:2}.chart-container .dataset-units path{fill:none;stroke-opacity:1;stroke-width:2px}.chart-container .dataset-path{stroke-width:2px}.chart-container .path-group path{fill:none;stroke-opacity:1;stroke-width:2px}.chart-container line.dashed{stroke-dasharray:5,3}.chart-container .axis-line .specific-value{text-anchor:start}.chart-container .axis-line .y-line{text-anchor:end}.chart-container .axis-line .x-line{text-anchor:middle}.chart-container .legend-dataset-text{fill:#6c7680;font-weight:600}.graph-svg-tip{position:absolute;z-index:99999;padding:10px;font-size:12px;color:#959da5;text-align:center;background:rgba(0,0,0,.8);border-radius:3px}.graph-svg-tip ul{padding-left:0;display:flex}.graph-svg-tip ol{padding-left:0;display:flex}.graph-svg-tip ul.data-point-list li{min-width:90px;flex:1;font-weight:600}.graph-svg-tip strong{color:#dfe2e5;font-weight:600}.graph-svg-tip .svg-pointer{position:absolute;height:5px;margin:0 0 0 -5px;content:' ';border:5px solid transparent;border-top-color:rgba(0,0,0,.8)}.graph-svg-tip.comparison{padding:0;text-align:left;pointer-events:none}.graph-svg-tip.comparison .title{display:block;padding:10px;margin:0;font-weight:600;line-height:1;pointer-events:none}.graph-svg-tip.comparison ul{margin:0;white-space:nowrap;list-style:none}.graph-svg-tip.comparison li{display:inline-block;padding:5px 10px}"});s.insertBefore(i,s.firstChild);let a=t.create("div");return a.appendChild(s),a.innerHTML}(this.svg);!function(t,e){var s=document.createElement("a");s.style="display: none";var i=new Blob(e,{type:"image/svg+xml; charset=utf-8"}),a=window.URL.createObjectURL(i);s.href=a,s.download=t,document.body.appendChild(s),s.click(),setTimeout((function(){document.body.removeChild(s),window.URL.revokeObjectURL(a)}),300)}(this.title||"Chart",[e])}}class V extends B{constructor(t,e){super(t,e)}configure(t){super.configure(t),this.config.formatTooltipY=(t.tooltipOptions||{}).formatTooltipY,this.config.maxSlices=t.maxSlices||20,this.config.maxLegendPoints=t.maxLegendPoints||20}calc(){let t=this.state,e=this.config.maxSlices;t.sliceTotals=[];let s=this.data.labels.map(((t,e)=>{let s=0;return this.data.datasets.map((t=>{s+=t.values[e]})),[s,t]})).filter((t=>t[0]>=0)),i=s;if(s.length>e){s.sort(((t,e)=>e[0]-t[0])),i=s.slice(0,e-1);let t=s.slice(e-1),a=0;t.map((t=>{a+=t[0]})),i.push([a,"Rest"]),this.colors[e-1]="grey"}t.labels=[],i.map((e=>{t.sliceTotals.push(function(t){return Number(Math.round(t+"e4")+"e-4")}(e[0])),t.labels.push(e[1])})),t.grandTotal=t.sliceTotals.reduce(((t,e)=>t+e),0),this.center={x:this.width/2,y:this.height/2}}renderLegend(){let t=this.state;this.legendArea.textContent="",this.legendTotals=t.sliceTotals.slice(0,this.config.maxLegendPoints);let e=0,s=0;this.legendTotals.map(((i,a)=>{let n=150,o=Math.floor((this.width-r(this.measures))/n);this.legendTotals.length<o&&(n=this.width/this.legendTotals.length),e>o&&(e=0,s+=20);let l=n*e+5,h=this.config.truncateLegends?x(t.labels[a],n/10):t.labels[a],d=this.config.formatTooltipY?this.config.formatTooltipY(i):i,c=function(t,e,s,i="none",a,n=!1){let r={className:"legend-dot",cx:0,cy:0,r:s,fill:i},o=M("text",{className:"legend-dataset-text",x:0,y:0,dx:"10px",dy:10/3+"px","font-size":"12px","text-anchor":"start",fill:"#555b51",innerHTML:a=n?x(a,15):a}),l=M("g",{transform:`translate(${t}, ${e})`});return l.appendChild(M("circle",r)),l.appendChild(o),l}(l,s,5,this.colors[a],`${h}: ${d}`,!1);this.legendArea.appendChild(c),e++}))}}const U=["January","February","March","April","May","June","July","August","September","October","November","December"],G=["Sun","Mon","Tue","Wed","Thu","Fri","Sat"];function q(t){let e=new Date(t);return e.setMinutes(e.getMinutes()-e.getTimezoneOffset()),e}function X(t){let e=t.getDate(),s=t.getMonth()+1;return[t.getFullYear(),(s>9?"":"0")+s,(e>9?"":"0")+e].join("-")}function _(t){return new Date(t.getTime())}function J(t,e){let s=tt(t);return Math.ceil(function(t,e){let s=864e5;return(q(e)-q(t))/s}(s,e)/7)}function K(t,e){return t.getMonth()===e.getMonth()&&t.getFullYear()===e.getFullYear()}function Q(t,e=!1){let s=U[t];return e?s.slice(0,3):s}function Z(t,e){return new Date(e,t+1,0)}function tt(t){let e=_(t);const s=e.getDay();return 0!==s&&et(e,-1*s),e}function et(t,e){t.setDate(t.getDate()+e)}class st{constructor({layerClass:t="",layerTransform:e="",constants:s,getData:i,makeElements:a,animateElements:n}){this.layerTransform=e,this.constants=s,this.makeElements=a,this.getData=i,this.animateElements=n,this.store=[],this.labels=[],this.layerClass=t,this.layerClass="function"==typeof this.layerClass?this.layerClass():this.layerClass,this.refresh()}refresh(t){this.data=t||this.getData()}setup(t){this.layer=P(this.layerClass,this.layerTransform,t)}make(){this.render(this.data),this.oldData=this.data}render(t){this.store=this.makeElements(t),this.layer.textContent="",this.store.forEach((t=>{this.layer.appendChild(t)})),this.labels.forEach((t=>{this.layer.appendChild(t)}))}update(t=!0){this.refresh();let e=[];return t&&(e=this.animateElements(this.data)||[]),e}}let it={donutSlices:{layerClass:"donut-slices",makeElements:t=>t.sliceStrings.map(((e,s)=>{let i=C(e,"donut-path",t.colors[s],"none",t.strokeWidth);return i.style.transition="transform .3s;",i})),animateElements(t){return this.store.map(((e,s)=>R(e,t.sliceStrings[s])))}},pieSlices:{layerClass:"pie-slices",makeElements:t=>t.sliceStrings.map(((e,s)=>{let i=C(e,"pie-path","none",t.colors[s]);return i.style.transition="transform .3s;",i})),animateElements(t){return this.store.map(((e,s)=>R(e,t.sliceStrings[s])))}},percentageBars:{layerClass:"percentage-bars",makeElements(t){return t.xPositions.map(((e,s)=>function(t,e,s,i,a=2,n="none"){return M("rect",{className:"percentage-bar",x:t,y:e,width:s,height:i,fill:n,styles:{stroke:k(n,-25),"stroke-dasharray":`0, ${i+s}, ${s}, ${i}`,"stroke-width":a}})}(e,0,t.widths[s],this.constants.barHeight,this.constants.barDepth,t.colors[s])))},animateElements(t){if(t)return[]}},yAxis:{layerClass:"y axis",makeElements(t){return t.positions.map(((e,s)=>function(t,e,s,i={}){g(t)||(t=0),i.pos||(i.pos="left"),i.offset||(i.offset=0),i.mode||(i.mode="span"),i.stroke||(i.stroke="#dadada"),i.className||(i.className="");let a=-6,n="span"===i.mode?s+6:0;return"tick"===i.mode&&"right"===i.pos&&(a=s+6,n=s),a+=i.offset,n+=i.offset,O(t,e,a,n,{stroke:i.stroke,className:i.className,lineType:i.lineType,shortenNumbers:i.shortenNumbers})}(e,t.labels[s],this.constants.width,{mode:this.constants.mode,pos:this.constants.pos,shortenNumbers:this.constants.shortenNumbers})))},animateElements(t){let e=t.positions,s=t.labels,i=this.oldData.positions,a=this.oldData.labels;return[i,e]=b(i,e),[a,s]=b(a,s),this.render({positions:i,labels:s}),this.store.map(((t,s)=>F(t,e[s],i[s])))}},xAxis:{layerClass:"x axis",makeElements(t){return t.positions.map(((e,s)=>S(e,t.calcLabels[s],this.constants.height,{mode:this.constants.mode,pos:this.constants.pos})))},animateElements(t){let e=t.positions,s=t.calcLabels,i=this.oldData.positions,a=this.oldData.calcLabels;return[i,e]=b(i,e),[a,s]=b(a,s),this.render({positions:i,calcLabels:s}),this.store.map(((t,s)=>function(t,e,s){return H(t,[s,0],[e,0],350)}(t,e[s],i[s])))}},yMarkers:{layerClass:"y-markers",makeElements(t){return t.map((t=>function(t,e,s,i={}){i.labelPos||(i.labelPos="right");let a=M("text",{className:"chart-label",x:"left"===i.labelPos?4:s-u(e,5)-4,y:0,dy:"-5px","font-size":"10px","text-anchor":"start",innerHTML:e+""}),n=O(t,"",0,s,{stroke:i.stroke||"#dadada",className:i.className||"",lineType:i.lineType});return n.appendChild(a),n}(t.position,t.label,this.constants.width,{labelPos:t.options.labelPos,mode:"span",lineType:"dashed"})))},animateElements(t){[this.oldData,t]=b(this.oldData,t);let e=t.map((t=>t.position)),s=t.map((t=>t.label)),i=t.map((t=>t.options)),a=this.oldData.map((t=>t.position));return this.render(a.map(((t,e)=>({position:a[e],label:s[e],options:i[e]})))),this.store.map(((t,s)=>F(t,e[s],a[s])))}},yRegions:{layerClass:"y-regions",makeElements(t){return t.map((t=>function(t,e,s,i,a={}){let n=t-e,r=M("rect",{className:"bar mini",styles:{fill:"rgba(228, 234, 239, 0.49)",stroke:"#dadada","stroke-dasharray":`${s}, ${n}`},x:0,y:0,width:s,height:n});a.labelPos||(a.labelPos="right");let o=M("text",{className:"chart-label",x:"left"===a.labelPos?4:s-u(i+"",4.5)-4,y:0,dy:"-5px","font-size":"10px","text-anchor":"start",innerHTML:i+""}),l=M("g",{transform:`translate(0, ${e})`});return l.appendChild(r),l.appendChild(o),l}(t.startPos,t.endPos,this.constants.width,t.label,{labelPos:t.options.labelPos})))},animateElements(t){[this.oldData,t]=b(this.oldData,t);let e=t.map((t=>t.endPos)),s=t.map((t=>t.label)),i=t.map((t=>t.startPos)),a=t.map((t=>t.options)),n=this.oldData.map((t=>t.endPos)),r=this.oldData.map((t=>t.startPos));this.render(n.map(((t,e)=>({startPos:r[e],endPos:n[e],label:s[e],options:a[e]}))));let o=[];return this.store.map(((t,s)=>{o=o.concat(function(t,e,s,i){let a=e-s,n=t.childNodes[0],r=n.getAttribute("width");return[[n,{height:a,"stroke-dasharray":`${r}, ${a}`},350,"easein"],H(t,[0,i],[0,s],350)]}(t,i[s],e[s],n[s]))})),o}},heatDomain:{layerClass:function(){return"heat-domain domain-"+this.constants.index},makeElements(t){let{index:e,colWidth:s,rowHeight:i,squareSize:a,radius:n,xTranslate:r}=this.constants,o=r,l=0;return this.serializedSubDomains=[],t.cols.map(((t,r)=>{1===r&&this.labels.push(E("domain-name",o,-12,Q(e,!0).toUpperCase(),{fontSize:9})),t.map(((t,e)=>{if(t.fill){let s={"data-date":t.yyyyMmDd,"data-value":t.dataValue,"data-day":e},i=N("day",o,l,a,n,t.fill,s);this.serializedSubDomains.push(i)}l+=i})),l=0,o+=s})),this.serializedSubDomains},animateElements(t){if(t)return[]}},barGraph:{layerClass:function(){return"dataset-units dataset-bars dataset-"+this.constants.index},makeElements(t){let e=this.constants;return this.unitType="bar",this.units=t.yPositions.map(((s,i)=>function(t,e,s,i,a="",n=0,r=0,o={}){let[l,h]=y(e,o.zeroLine);h-=r,0===l&&(l=o.minHeight,h-=o.minHeight),g(t)||(t=0),g(h)||(h=0),g(l,!0)||(l=0),g(s,!0)||(s=0);let d=M("rect",{className:"bar mini",style:`fill: ${i}`,"data-point-index":n,x:t,y:h,width:s,height:l});if((a+="")||a.length){d.setAttribute("y",0),d.setAttribute("x",0);let e=M("text",{className:"data-point-value",x:s/2,y:0,dy:"-5px","font-size":"10px","text-anchor":"middle",innerHTML:a}),i=M("g",{"data-point-index":n,transform:`translate(${t}, ${h})`});return i.appendChild(d),i.appendChild(e),i}return d}(t.xPositions[i],s,t.barWidth,e.color,t.labels[i],i,t.offsets[i],{zeroLine:t.zeroLine,barsWidth:t.barsWidth,minHeight:e.minHeight}))),this.units},animateElements(t){let e=t.xPositions,s=t.yPositions,i=t.offsets,a=t.labels,n=this.oldData.xPositions,r=this.oldData.yPositions,o=this.oldData.offsets,l=this.oldData.labels;[n,e]=b(n,e),[r,s]=b(r,s),[o,i]=b(o,i),[l,a]=b(l,a),this.render({xPositions:n,yPositions:r,offsets:o,labels:a,zeroLine:this.oldData.zeroLine,barsWidth:this.oldData.barsWidth,barWidth:this.oldData.barWidth});let h=[];return this.store.map(((a,n)=>{h=h.concat(function(t,e,s,i,a=0,n={}){let[r,o]=y(s,n.zeroLine);if(o-=a,"rect"!==t.nodeName){let s=[t.childNodes[0],{width:i,height:r},350,"easein"],a=t.getAttribute("transform").split("(")[1].slice(0,-1);return[s,H(t,a,[e,o],350)]}return[[t,{width:i,height:r,x:e,y:o},350,"easein"]]}(a,e[n],s[n],t.barWidth,i[n],{zeroLine:t.zeroLine}))})),h}},lineGraph:{layerClass:function(){return"dataset-units dataset-line dataset-"+this.constants.index},makeElements(t){let e=this.constants;return this.unitType="dot",this.paths={},e.hideLine||(this.paths=function(t,e,s,i={},a={}){let n=e.map(((e,s)=>t[s]+","+e)).join("L");i.spline&&(n=v(t,e));let r=C("M"+n,"line-graph-path",s);if(i.heatline){let t=$(a.svgDefs,s);r.style.stroke=`url(#${t})`}let o={path:r};if(i.regionFill){let e=$(a.svgDefs,s,!0),i=`M${t[0]},${a.zeroLine}L`+n+`L${t.slice(-1)[0]},${a.zeroLine}`;o.region=C(i,"region-fill","none",`url(#${e})`)}return o}(t.xPositions,t.yPositions,e.color,{heatline:e.heatline,regionFill:e.regionFill,spline:e.spline},{svgDefs:e.svgDefs,zeroLine:t.zeroLine})),this.units=[],e.hideDots||(this.units=t.yPositions.map(((s,i)=>function(t,e,s,i,a="",n=0){let r=M("circle",{style:`fill: ${i}`,"data-point-index":n,cx:t,cy:e,r:s});if((a+="")||a.length){r.setAttribute("cy",0),r.setAttribute("cx",0);let i=M("text",{className:"data-point-value",x:0,y:0,dy:-5-s+"px","font-size":"10px","text-anchor":"middle",innerHTML:a}),o=M("g",{"data-point-index":n,transform:`translate(${t}, ${e})`});return o.appendChild(r),o.appendChild(i),o}return r}(t.xPositions[i],s,t.radius,e.color,e.valuesOverPoints?t.values[i]:"",i)))),Object.values(this.paths).concat(this.units)},animateElements(t){let e=t.xPositions,s=t.yPositions,i=t.values,a=this.oldData.xPositions,n=this.oldData.yPositions,r=this.oldData.values;[a,e]=b(a,e),[n,s]=b(n,s),[r,i]=b(r,i),this.render({xPositions:a,yPositions:n,values:i,zeroLine:this.oldData.zeroLine,radius:this.oldData.radius});let o=[];return Object.keys(this.paths).length&&(o=o.concat(function(t,e,s,i,a){let n=[],r=s.map(((t,s)=>e[s]+","+t)).join("L");a&&(r=v(e,s));const o=[t.path,{d:"M"+r},350,"easein"];if(n.push(o),t.region){let s=`${e[0]},${i}L`,a=`L${e.slice(-1)[0]}, ${i}`;const o=[t.region,{d:"M"+s+r+a},350,"easein"];n.push(o)}return n}(this.paths,e,s,t.zeroLine,this.constants.spline))),this.units.length&&this.units.map(((t,i)=>{o=o.concat(function(t,e,s){if("circle"!==t.nodeName){let i=t.getAttribute("transform").split("(")[1].slice(0,-1);return[H(t,i,[e,s],350)]}return[[t,{cx:e,cy:s},350,"easein"]]}(t,e[i],s[i]))})),o}}};function at(t,e,s){let i=Object.keys(it).filter((e=>t.includes(e))),a=it[i[0]];return Object.assign(a,{constants:e,getData:s}),new st(a)}function nt(t){if(0===t)return[0,0];if(isNaN(t))return{mantissa:-6755399441055744,exponent:972};var e=t>0?1:-1;if(!isFinite(t))return{mantissa:4503599627370496*e,exponent:972};t=Math.abs(t);var s=Math.floor(Math.log10(t));return[e*(t/Math.pow(10,s)),s]}function rt(t,e=0){let[s,i]=nt(t),a=e?e/Math.pow(10,i):0;s=s.toFixed(6);let n=function(t,e=0){let s=Math.ceil(t),i=Math.floor(e),a=s-i,n=a,r=1;a>5&&(a%2!=0&&(s++,a=s-i),n=a/2,r=2),a<=2&&(n=4,r=a/n),0===a&&(n=5,r=1);let o=[];for(var l=0;l<=n;l++)o.push(i+r*l);return o}(s,a);return n=n.map((t=>t*Math.pow(10,i))),n}function ot(t){return t[1]-t[0]}function lt(t,e){return c(e.zeroLine-t*e.scaleMultiplier)}class ht extends B{constructor(t,e){super(t,e),this.barOptions=e.barOptions||{},this.lineOptions=e.lineOptions||{},this.type=e.type||"line",this.init=1,this.setup()}setMeasures(){this.data.datasets.length<=1&&(this.config.showLegend=0,this.measures.paddings.bottom=30)}configure(t){super.configure(t),t.axisOptions=t.axisOptions||{},t.tooltipOptions=t.tooltipOptions||{},this.config.xAxisMode=t.axisOptions.xAxisMode||"span",this.config.yAxisMode=t.axisOptions.yAxisMode||"span",this.config.xIsSeries=t.axisOptions.xIsSeries||0,this.config.shortenYAxisNumbers=t.axisOptions.shortenYAxisNumbers||0,this.config.formatTooltipX=t.tooltipOptions.formatTooltipX,this.config.formatTooltipY=t.tooltipOptions.formatTooltipY,this.config.valuesOverPoints=t.valuesOverPoints}prepareData(t=this.data){return function(t,e){t.labels=t.labels||[];let s=t.labels.length,i=t.datasets,a=new Array(s).fill(0);return i||(i=[{values:a}]),i.map((t=>{if(t.values){let e=t.values;e=e.map((t=>isNaN(t)?0:t)),e=e.length>s?e.slice(0,s):p(e,s-e.length,0),t.values=e}else t.values=a;t.chartType||(t.chartType=e)})),t.yRegions&&t.yRegions.map((t=>{t.end<t.start&&([t.start,t.end]=[t.end,t.start])})),t}(t,this.type)}prepareFirstData(t=this.data){return function(t){let e=t.labels.length,s=new Array(e).fill(0),i={labels:t.labels.slice(0,-1),datasets:t.datasets.map((t=>({name:"",values:s.slice(0,-1),chartType:t.chartType})))};return t.yMarkers&&(i.yMarkers=[{value:0,label:""}]),t.yRegions&&(i.yRegions=[{start:0,end:0,label:""}]),i}(t)}calc(t=!1){this.calcXPositions(),t||this.calcYAxisParameters(this.getAllYValues(),"line"===this.type),this.makeDataByIndex()}calcXPositions(){let t=this.state,e=this.data.labels;t.datasetLength=e.length,t.unitWidth=this.width/t.datasetLength,t.xOffset=t.unitWidth/2,t.xAxis={labels:e,positions:e.map(((e,s)=>c(t.xOffset+s*t.unitWidth)))}}calcYAxisParameters(t,e="false"){const s=function(t,e=!1){let s=Math.max(...t),i=Math.min(...t),a=[];function n(t,e){let s=rt(t),i=s[1]-s[0],a=0;for(var n=1;a<e;n++)a+=i,s.unshift(-1*a);return s}if(s>=0&&i>=0)nt(s)[1],a=e?rt(s,i):rt(s);else if(s>0&&i<0){let t=Math.abs(i);s>=t?(nt(s)[1],a=n(s,t)):(nt(t)[1],a=n(t,s).reverse().map((t=>-1*t)))}else if(s<=0&&i<=0){let t=Math.abs(i),n=Math.abs(s);nt(t)[1],a=e?rt(t,n):rt(t),a=a.reverse().map((t=>-1*t))}return a}(t,e),i=this.height/((a=s)[a.length-1]-a[0]);var a;const n=ot(s)*i,r=this.height-function(t){let e,s=ot(t);e=t.indexOf(0)>=0?t.indexOf(0):t[0]>0?-1*t[0]/s:-1*t[t.length-1]/s+(t.length-1);return e}(s)*n;this.state.yAxis={labels:s,positions:s.map((t=>r-t*i)),scaleMultiplier:i,zeroLine:r},this.calcDatasetPoints(),this.calcYExtremes(),this.calcYRegions()}calcDatasetPoints(){let t=this.state,e=e=>e.map((e=>lt(e,t.yAxis)));t.datasets=this.data.datasets.map(((t,s)=>{let i=t.values,a=t.cumulativeYs||[];return{name:t.name&&t.name.replace(/<|>|&/g,(t=>"&"==t?"&amp;":"<"==t?"&lt;":"&gt;")),index:s,chartType:t.chartType,values:i,yPositions:e(i),cumulativeYs:a,cumulativeYPos:e(a)}}))}calcYExtremes(){let t=this.state;this.barOptions.stacked?t.yExtremes=t.datasets[t.datasets.length-1].cumulativeYPos:(t.yExtremes=new Array(t.datasetLength).fill(9999),t.datasets.map((e=>{e.yPositions.map(((e,s)=>{e<t.yExtremes[s]&&(t.yExtremes[s]=e)}))})))}calcYRegions(){let t=this.state;this.data.yMarkers&&(this.state.yMarkers=this.data.yMarkers.map((e=>(e.position=lt(e.value,t.yAxis),e.options||(e.options={}),e)))),this.data.yRegions&&(this.state.yRegions=this.data.yRegions.map((e=>(e.startPos=lt(e.start,t.yAxis),e.endPos=lt(e.end,t.yAxis),e.options||(e.options={}),e))))}getAllYValues(){let t="values";if(this.barOptions.stacked){t="cumulativeYs";let e=new Array(this.state.datasetLength).fill(0);this.data.datasets.map(((s,i)=>{let a=this.data.datasets[i].values;s[t]=e=e.map(((t,e)=>t+a[e]))}))}let e=this.data.datasets.map((e=>e[t]));return this.data.yMarkers&&e.push(this.data.yMarkers.map((t=>t.value))),this.data.yRegions&&this.data.yRegions.map((t=>{e.push([t.end,t.start])})),[].concat(...e)}setupComponents(){let t=[["yAxis",{mode:this.config.yAxisMode,width:this.width,shortenNumbers:this.config.shortenYAxisNumbers},function(){return this.state.yAxis}.bind(this)],["xAxis",{mode:this.config.xAxisMode,height:this.height},function(){let t=this.state;return t.xAxis.calcLabels=function(t,e=[],s=!0){let i=t/e.length;i<=0&&(i=1);let a,n=i/7;if(s){let t=Math.max(...e.map((t=>t.length)));a=Math.ceil(t/n)}return e.map(((t,e)=>((t+="").length>n&&(s?e%a!=0&&(t=""):t=n-3>0?t.slice(0,n-3)+" ...":t.slice(0,n)+".."),t)))}(this.width,t.xAxis.labels,this.config.xIsSeries),t.xAxis}.bind(this)],["yRegions",{width:this.width,pos:"right"},function(){return this.state.yRegions}.bind(this)]],e=this.state.datasets.filter((t=>"bar"===t.chartType)),s=this.state.datasets.filter((t=>"line"===t.chartType)),i=e.map((t=>{let s=t.index;return["barGraph-"+t.index,{index:s,color:this.colors[s],stacked:this.barOptions.stacked,valuesOverPoints:this.config.valuesOverPoints,minHeight:0*this.height},function(){let t=this.state,i=t.datasets[s],a=this.barOptions.stacked,n=this.barOptions.spaceRatio||.5,r=t.unitWidth*(1-n),o=r/(a?1:e.length),l=t.xAxis.positions.map((t=>t-r/2));a||(l=l.map((t=>t+o*s)));let h=new Array(t.datasetLength).fill("");this.config.valuesOverPoints&&(h=a&&i.index===t.datasets.length-1?i.cumulativeYs:i.values);let d=new Array(t.datasetLength).fill(0);return a&&(d=i.yPositions.map(((t,e)=>t-i.cumulativeYPos[e]))),{xPositions:l,yPositions:i.yPositions,offsets:d,labels:h,zeroLine:t.yAxis.zeroLine,barsWidth:r,barWidth:o}}.bind(this)]})),a=s.map((t=>{let e=t.index;return["lineGraph-"+t.index,{index:e,color:this.colors[e],svgDefs:this.svgDefs,heatline:this.lineOptions.heatline,regionFill:this.lineOptions.regionFill,spline:this.lineOptions.spline,hideDots:this.lineOptions.hideDots,hideLine:this.lineOptions.hideLine,valuesOverPoints:this.config.valuesOverPoints},function(){let t=this.state,s=t.datasets[e],i=t.yAxis.positions[0]<t.yAxis.zeroLine?t.yAxis.positions[0]:t.yAxis.zeroLine;return{xPositions:t.xAxis.positions,yPositions:s.yPositions,values:s.values,zeroLine:i,radius:this.lineOptions.dotSize||4}}.bind(this)]})),n=[["yMarkers",{width:this.width,pos:"right"},function(){return this.state.yMarkers}.bind(this)]];t=t.concat(i,a,n);let r=["yMarkers","yRegions"];this.dataUnitComponents=[],this.components=new Map(t.filter((t=>!r.includes(t[0])||this.state[t[0]])).map((t=>{let e=at(...t);return(t[0].includes("lineGraph")||t[0].includes("barGraph"))&&this.dataUnitComponents.push(e),[t[0],e]})))}makeDataByIndex(){this.dataByIndex={};let t=this.state,e=this.config.formatTooltipX,s=this.config.formatTooltipY;t.xAxis.labels.map(((i,a)=>{let n=this.state.datasets.map(((t,e)=>{let i=t.values[a];return{title:t.name,value:i,yPos:t.yPositions[a],color:this.colors[e],formatted:s?s(i):i}}));this.dataByIndex[a]={label:i,formattedLabel:e?e(i):i,xPos:t.xAxis.positions[a],values:n,yExtreme:t.yExtremes[a]}}))}bindTooltip(){this.container.addEventListener("mousemove",(t=>{let s=this.measures,n=e(this.container),r=t.pageX-n.left-a(s),o=t.pageY-n.top;o<this.height+i(s)&&o>i(s)?this.mapTooltipXPosition(r):this.tip.hideTip()}))}mapTooltipXPosition(t){let e=this.state;if(!e.yExtremes)return;let s=function(t,e,s=!1){let i=e.reduce((function(e,s){return Math.abs(s-t)<Math.abs(e-t)?s:e}),[]);return s?e.indexOf(i):i}(t,e.xAxis.positions,!0);if(s>=0){let t=this.dataByIndex[s];this.tip.setValues(t.xPos+this.tip.offset.x,t.yExtreme+this.tip.offset.y,{name:t.formattedLabel,value:""},t.values,s),this.tip.showTip()}}renderLegend(){let t=this.data;t.datasets.length>1&&(this.legendArea.textContent="",t.datasets.map(((t,e)=>{let s=function(t,e,s,i="none",a,n=!1){let r={className:"legend-bar",x:0,y:0,width:s,height:"2px",fill:i},o=M("text",{className:"legend-dataset-text",x:0,y:0,dy:"20px","font-size":"12px","text-anchor":"start",fill:"#555b51",innerHTML:a=n?x(a,15):a}),l=M("g",{transform:`translate(${t}, ${e})`});return l.appendChild(M("rect",r)),l.appendChild(o),l}(100*e,"0",100,this.colors[e],t.name,this.config.truncateLegends);this.legendArea.appendChild(s)})))}makeOverlay(){this.init?this.init=0:(this.overlayGuides&&this.overlayGuides.forEach((t=>{let e=t.overlay;e.parentNode.removeChild(e)})),this.overlayGuides=this.dataUnitComponents.map((t=>({type:t.unitType,overlay:void 0,units:t.units}))),void 0===this.state.currentIndex&&(this.state.currentIndex=this.state.datasetLength-1),this.overlayGuides.map((t=>{let e=t.units[this.state.currentIndex];t.overlay=z[t.type](e),this.drawArea.appendChild(t.overlay)})))}updateOverlayGuides(){this.overlayGuides&&this.overlayGuides.forEach((t=>{let e=t.overlay;e.parentNode.removeChild(e)}))}bindOverlay(){this.parent.addEventListener("data-select",(()=>{this.updateOverlay()}))}bindUnits(){this.dataUnitComponents.map((t=>{t.units.map((t=>{t.addEventListener("click",(()=>{let e=t.getAttribute("data-point-index");this.setCurrentDataPoint(e)}))}))})),this.tip.container.addEventListener("click",(()=>{let t=this.tip.container.getAttribute("data-point-index");this.setCurrentDataPoint(t)}))}updateOverlay(){this.overlayGuides.map((t=>{let e=t.units[this.state.currentIndex];W[t.type](e,t.overlay)}))}onLeftArrow(){this.setCurrentDataPoint(this.state.currentIndex-1)}onRightArrow(){this.setCurrentDataPoint(this.state.currentIndex+1)}getDataPoint(t=this.state.currentIndex){let e=this.state;return{index:t,label:e.xAxis.labels[t],values:e.datasets.map((e=>e.values[t]))}}setCurrentDataPoint(t){let e=this.state;(t=parseInt(t))<0&&(t=0),t>=e.xAxis.labels.length&&(t=e.xAxis.labels.length-1),t!==e.currentIndex&&(e.currentIndex=t,function(t,e,s){var i=document.createEvent("HTMLEvents");for(var a in i.initEvent(e,!0,!0),s)i[a]=s[a];t.dispatchEvent(i)}(this.parent,"data-select",this.getDataPoint()))}addDataPoint(t,e,s=this.state.datasetLength){super.addDataPoint(t,e,s),this.data.labels.splice(s,0,t),this.data.datasets.map(((t,i)=>{t.values.splice(s,0,e[i])})),this.update(this.data)}removeDataPoint(t=this.state.datasetLength-1){this.data.labels.length<=1||(super.removeDataPoint(t),this.data.labels.splice(t,1),this.data.datasets.map((e=>{e.values.splice(t,1)})),this.update(this.data))}updateDataset(t,e=0){this.data.datasets[e].values=t,this.update(this.data)}updateDatasets(t){this.data.datasets.map(((e,s)=>{t[s]&&(e.values=t[s])})),this.update(this.data)}}const dt={bar:ht,line:ht,percentage:class extends V{constructor(t,e){super(t,e),this.type="percentage",this.setup()}setMeasures(t){let e=this.measures;this.barOptions=t.barOptions||{};let s=this.barOptions;s.height=s.height||20,s.depth=s.depth||2,e.paddings.right=30,e.legendHeight=60,e.baseHeight=8*(s.height+.5*s.depth)}setupComponents(){let t=this.state,e=[["percentageBars",{barHeight:this.barOptions.height,barDepth:this.barOptions.depth},function(){return{xPositions:t.xPositions,widths:t.widths,colors:this.colors}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calc(){super.calc();let t=this.state;t.xPositions=[],t.widths=[];let e=0;t.sliceTotals.map((s=>{let i=this.width*s/t.grandTotal;t.widths.push(i),t.xPositions.push(e),e+=i}))}makeDataByIndex(){}bindTooltip(){let t=this.state;this.container.addEventListener("mousemove",(s=>{let i=this.components.get("percentageBars").store,a=s.target;if(i.includes(a)){let s=i.indexOf(a),n=e(this.container),r=e(a),o=r.left-n.left+parseInt(a.getAttribute("width"))/2,l=r.top-n.top,h=(this.formattedLabels&&this.formattedLabels.length>0?this.formattedLabels[s]:this.state.labels[s])+": ",d=t.sliceTotals[s]/t.grandTotal;this.tip.setValues(o,l,{name:h,value:(100*d).toFixed(1)+"%"}),this.tip.showTip()}}))}},heatmap:class extends B{constructor(t,e){super(t,e),this.type="heatmap",this.countLabel=e.countLabel||"";let s=["Sunday","Monday"],i=s.includes(e.startSubDomain)?e.startSubDomain:"Sunday";this.startSubDomainIndex=s.indexOf(i),this.setup()}setMeasures(t){let e=this.measures;this.discreteDomains=0===t.discreteDomains?0:1,e.paddings.top=36,e.paddings.bottom=0,e.legendHeight=24,e.baseHeight=84+n(e);let s=this.data,i=this.discreteDomains?12:0;this.independentWidth=12*(J(s.start,s.end)+i)+r(e)}updateWidth(){let t=this.discreteDomains?12:0,e=this.state.noOfWeeks?this.state.noOfWeeks:52;this.baseWidth=12*(e+t)+r(this.measures)}prepareData(t=this.data){if(t.start&&t.end&&t.start>t.end)throw new Error("Start date cannot be greater than end date.");if(t.start||(t.start=new Date,t.start.setFullYear(t.start.getFullYear()-1)),t.end||(t.end=new Date),t.dataPoints=t.dataPoints||{},parseInt(Object.keys(t.dataPoints)[0])>1e5){let e={};Object.keys(t.dataPoints).forEach((s=>{let i=new Date(1e3*s);e[X(i)]=t.dataPoints[s]})),t.dataPoints=e}return t}calc(){let t=this.state;t.start=_(this.data.start),t.end=_(this.data.end),t.firstWeekStart=_(t.start),t.noOfWeeks=J(t.start,t.end),t.distribution=function(t,e){let s=Math.max(...t),i=1/(e-1),a=[];for(var n=0;n<e;n++){let t=s*(i*n);a.push(t)}return a}(Object.values(this.data.dataPoints),5),t.domainConfigs=this.getDomains()}setupComponents(){let t=this.state,e=this.discreteDomains?0:1,s=t.domainConfigs.map(((s,i)=>["heatDomain",{index:s.index,colWidth:12,rowHeight:12,squareSize:10,radius:this.rawChartArgs.radius||0,xTranslate:12*t.domainConfigs.filter(((t,e)=>e<i)).map((t=>t.cols.length-e)).reduce(((t,e)=>t+e),0)},function(){return t.domainConfigs[i]}.bind(this)]));this.components=new Map(s.map(((t,e)=>{let s=at(...t);return[t[0]+"-"+e,s]})));let i=0;G.forEach(((t,e)=>{if([1,3,5].includes(e)){let e=E("subdomain-name",-6,i,t,{fontSize:10,dy:8,textAnchor:"end"});this.drawArea.appendChild(e)}i+=12}))}update(t){t||console.error("No data to update."),this.data=this.prepareData(t),this.draw(),this.bindTooltip()}bindTooltip(){this.container.addEventListener("mousemove",(t=>{this.components.forEach((e=>{let s=e.store,i=t.target;if(s.includes(i)){let e=i.getAttribute("data-value"),s=i.getAttribute("data-date").split("-"),a=Q(parseInt(s[1])-1,!0),n=this.container.getBoundingClientRect(),r=i.getBoundingClientRect(),o=parseInt(t.target.getAttribute("width")),l=r.left-n.left+o/2,h=r.top-n.top,d=e+" "+this.countLabel,c=" on "+a+" "+s[0]+", "+s[2];this.tip.setValues(l,h,{name:c,value:d,valueFirst:1},[]),this.tip.showTip()}}))}))}renderLegend(){this.legendArea.textContent="";let t=0,e=this.rawChartArgs.radius||0,s=E("subdomain-name",t,12,"Less",{fontSize:11,dy:9});t=30,this.legendArea.appendChild(s),this.colors.slice(0,5).map(((s,i)=>{const a=N("heatmap-legend-unit",t+15*i,12,10,e,s);this.legendArea.appendChild(a)}));let i=E("subdomain-name",t+75+3,12,"More",{fontSize:11,dy:9});this.legendArea.appendChild(i)}getDomains(){let t=this.state;const[e,s]=[t.start.getMonth(),t.start.getFullYear()],[i,a]=[t.end.getMonth(),t.end.getFullYear()],n=i-e+1+12*(a-s);let r=[],o=_(t.start);for(var l=0;l<n;l++){let e=t.end;if(!K(o,t.end)){let[t,s]=[o.getMonth(),o.getFullYear()];e=Z(t,s)}r.push(this.getDomainConfig(o,e)),et(e,1),o=e}return r}getDomainConfig(t,e=""){let[s,i]=[t.getMonth(),t.getFullYear()],a=tt(t),n={index:s,cols:[]};et(e=_(e)||Z(s,i),1);let r,o=J(a,e),l=[];for(var h=0;h<o;h++)r=this.getCol(a,s),l.push(r),a=new Date(r[6].yyyyMmDd),et(a,1);return void 0!==r[6].dataValue&&(et(a,1),l.push(this.getCol(a,s,!0))),n.cols=l,n}getCol(t,e,s=!1){let i=this.state,a=_(t),n=[];for(var r=0;r<7;r++,et(a,1)){let t={},r=a>=i.start&&a<=i.end;s||a.getMonth()!==e||!r?t.yyyyMmDd=X(a):t=this.getSubDomainConfig(a),n.push(t)}return n}getSubDomainConfig(t){let e=X(t),s=this.data.dataPoints[e];var i,a;return{yyyyMmDd:e,dataValue:s||0,fill:this.colors[(i=s,a=this.state.distribution,a.filter((t=>t<i)).length)]}}},pie:class extends V{constructor(t,e){super(t,e),this.type="pie",this.initTimeout=0,this.init=1,this.setup()}configure(t){super.configure(t),this.mouseMove=this.mouseMove.bind(this),this.mouseLeave=this.mouseLeave.bind(this),this.hoverRadio=t.hoverRadio||.1,this.config.startAngle=t.startAngle||0,this.clockWise=t.clockWise||!1}calc(){super.calc();let t=this.state;this.radius=this.height>this.width?this.center.x:this.center.y;const{radius:e,clockWise:s}=this,i=t.slicesProperties||[];t.sliceStrings=[],t.slicesProperties=[];let a=180-this.config.startAngle;t.sliceTotals.map(((n,r)=>{const o=a,l=n/t.grandTotal*360,h=l>180?1:0,d=s?-l:l,c=a+=d,p=m(o,e),u=m(c,e),g=this.init&&i[r];let f,y;this.init?(f=g?g.startPosition:p,y=g?g.endPosition:p):(f=p,y=u);const b=360===l?function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h,d]=[s.x+e.x,2*s.y,s.y+e.y];return`M${s.x} ${s.y}\n\t\tL${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h} z\n\t\tL${r} ${h}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${d} z`}(f,y,this.center,this.radius,s,h):function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h]=[s.x+e.x,s.y+e.y];return`M${s.x} ${s.y}\n\t\tL${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h} z`}(f,y,this.center,this.radius,s,h);t.sliceStrings.push(b),t.slicesProperties.push({startPosition:p,endPosition:u,value:n,total:t.grandTotal,startAngle:o,endAngle:c,angle:d})})),this.init=0}setupComponents(){let t=this.state,e=[["pieSlices",{},function(){return{sliceStrings:t.sliceStrings,colors:this.colors}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calTranslateByAngle(t){const{radius:e,hoverRadio:s}=this,i=m(t.startAngle+t.angle/2,e);return`translate3d(${i.x*s}px,${i.y*s}px,0)`}hoverSlice(t,s,i,a){if(!t)return;const n=this.colors[s];if(i){Y(t,this.calTranslateByAngle(this.state.slicesProperties[s])),t.style.fill=k(n,50);let i=e(this.svg),r=a.pageX-i.left+10,o=a.pageY-i.top-10,l=(this.formatted_labels&&this.formatted_labels.length>0?this.formatted_labels[s]:this.state.labels[s])+": ",h=(100*this.state.sliceTotals[s]/this.state.grandTotal).toFixed(1);this.tip.setValues(r,o,{name:l,value:h+"%"}),this.tip.showTip()}else Y(t,"translate3d(0,0,0)"),this.tip.hideTip(),t.style.fill=n}bindTooltip(){this.container.addEventListener("mousemove",this.mouseMove),this.container.addEventListener("mouseleave",this.mouseLeave)}mouseMove(t){const e=t.target;let s=this.components.get("pieSlices").store,i=this.curActiveSliceIndex,a=this.curActiveSlice;if(s.includes(e)){let n=s.indexOf(e);this.hoverSlice(a,i,!1),this.curActiveSlice=e,this.curActiveSliceIndex=n,this.hoverSlice(e,n,!0,t)}else this.mouseLeave()}mouseLeave(){this.hoverSlice(this.curActiveSlice,this.curActiveSliceIndex,!1)}},donut:class extends V{constructor(t,e){super(t,e),this.type="donut",this.initTimeout=0,this.init=1,this.setup()}configure(t){super.configure(t),this.mouseMove=this.mouseMove.bind(this),this.mouseLeave=this.mouseLeave.bind(this),this.hoverRadio=t.hoverRadio||.1,this.config.startAngle=t.startAngle||0,this.clockWise=t.clockWise||!1,this.strokeWidth=t.strokeWidth||30}calc(){super.calc();let t=this.state;this.radius=this.height>this.width?this.center.x-this.strokeWidth/2:this.center.y-this.strokeWidth/2;const{radius:e,clockWise:s}=this,i=t.slicesProperties||[];t.sliceStrings=[],t.slicesProperties=[];let a=180-this.config.startAngle;t.sliceTotals.map(((n,r)=>{const o=a,l=n/t.grandTotal*360,h=l>180?1:0,d=s?-l:l,c=a+=d,p=m(o,e),u=m(c,e),g=this.init&&i[r];let f,y;this.init?(f=g?g.startPosition:p,y=g?g.endPosition:p):(f=p,y=u);const b=360===l?function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h,d]=[s.x+e.x,2*i+o,s.y+t.y];return`M${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h}\n\t\tM${r} ${h}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${d}`}(f,y,this.center,this.radius,this.clockWise,h):function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h]=[s.x+e.x,s.y+e.y];return`M${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h}`}(f,y,this.center,this.radius,this.clockWise,h);t.sliceStrings.push(b),t.slicesProperties.push({startPosition:p,endPosition:u,value:n,total:t.grandTotal,startAngle:o,endAngle:c,angle:d})})),this.init=0}setupComponents(){let t=this.state,e=[["donutSlices",{},function(){return{sliceStrings:t.sliceStrings,colors:this.colors,strokeWidth:this.strokeWidth}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calTranslateByAngle(t){const{radius:e,hoverRadio:s}=this,i=m(t.startAngle+t.angle/2,e);return`translate3d(${i.x*s}px,${i.y*s}px,0)`}hoverSlice(t,s,i,a){if(!t)return;const n=this.colors[s];if(i){Y(t,this.calTranslateByAngle(this.state.slicesProperties[s])),t.style.stroke=k(n,50);let i=e(this.svg),r=a.pageX-i.left+10,o=a.pageY-i.top-10,l=(this.formatted_labels&&this.formatted_labels.length>0?this.formatted_labels[s]:this.state.labels[s])+": ",h=(100*this.state.sliceTotals[s]/this.state.grandTotal).toFixed(1);this.tip.setValues(r,o,{name:l,value:h+"%"}),this.tip.showTip()}else Y(t,"translate3d(0,0,0)"),this.tip.hideTip(),t.style.stroke=n}bindTooltip(){this.container.addEventListener("mousemove",this.mouseMove),this.container.addEventListener("mouseleave",this.mouseLeave)}mouseMove(t){const e=t.target;let s=this.components.get("donutSlices").store,i=this.curActiveSliceIndex,a=this.curActiveSlice;if(s.includes(e)){let n=s.indexOf(e);this.hoverSlice(a,i,!1),this.curActiveSlice=e,this.curActiveSliceIndex=n,this.hoverSlice(e,n,!0,t)}else this.mouseLeave()}mouseLeave(){this.hoverSlice(this.curActiveSlice,this.curActiveSliceIndex,!1)}}};class ct{constructor(t,e){return function(t="line",e,s){return"axis-mixed"===t?(s.type="line",new ht(e,s)):dt[t]?new dt[t](e,s):void console.error("Undefined chart type: "+t)}(e.type,t,e)}}const pt=document.getElementById("main"),ut=JSON.parse(document.querySelector(".data-json").innerHTML);let mt;const gt=(t,e)=>{const s={Fast:[],Average:[],Slow:[]},i=[];return Object.keys(ut).forEach((i=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((a=>{ut[i][a]&&ut[i][a].forEach((i=>{i&&i.client&&i.client===e&&i.platform&&i.platform.toLowerCase()===t.toLowerCase()&&(s.Slow.push(i.slow),s.Average.push(i.avg),s.Fast.push(i.fast))}))}))})),["Slow","Average","Fast"].forEach((t=>{i.push({name:t,chartType:"bar",values:s[t]})})),i},ft=()=>{const t=[];return Object.keys(ut).forEach((e=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((s=>{ut[e][s]&&t.push(`${s}/'${e.substr(2,2)}`)}))})),t},yt={yRegions:[{label:"%",start:0,end:100}],data:{labels:ft()},barOptions:{spaceRatio:.001,stacked:!0},height:400,colors:["red","yellow","green","black"],axisOptions:{xAxisMode:"tick",yAxisMode:"span"}},bt=(t,e)=>{mt?mt.update({labels:ft(),datasets:gt(t,e)}):(yt.data.datasets=gt(t,e),mt=new ct(pt,yt))},xt=(()=>{const t=[];return Object.keys(ut).forEach((e=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((s=>{ut[e][s]&&ut[e][s].forEach((e=>{e.platform&&!t.includes(e.platform.toLowerCase())&&t.push(e.platform.toLowerCase())}))}))})),t})(),vt=document.createElement("select"),wt=document.createElement("select");wt.innerHTML="<option value='mobile'>Mobile</option><option value='desktop'>Desktop</option>",xt.sort().forEach((t=>{const e=document.createElement("option");var s;e.text=(s=t)&&s[0].toUpperCase()+s.slice(1),e.value=t,vt.appendChild(e)})),vt.addEventListener("change",(t=>{bt(t.target.value,wt.value)})),wt.addEventListener("change",(t=>{bt(vt.value,t.target.value)})),pt.insertAdjacentElement("beforebegin",vt),pt.insertAdjacentElement("beforebegin",wt),bt(xt[0],"mobile");
+function t(t,e){return"string"==typeof t?(e||document).querySelector(t):t||null}function e(t){let e=t.getBoundingClientRect();return{top:e.top+(document.documentElement.scrollTop||document.body.scrollTop),left:e.left+(document.documentElement.scrollLeft||document.body.scrollLeft)}}t.create=(e,s)=>{var i=document.createElement(e);for(var a in s){var n=s[a];if("inside"===a)t(n).appendChild(i);else if("around"===a){var r=t(n);r.parentNode.insertBefore(i,r),i.appendChild(r)}else"styles"===a?"object"==typeof n&&Object.keys(n).map((t=>{i.style[t]=n[t]})):a in i?i[a]=n:i.setAttribute(a,n)}return i};const s={margins:{top:10,bottom:10,left:20,right:20},paddings:{top:20,bottom:40,left:30,right:10},baseHeight:240,titleHeight:20,legendHeight:30,titleFontSize:12};function i(t){return t.titleHeight+t.margins.top+t.paddings.top}function a(t){return t.margins.left+t.paddings.left}function n(t){return t.margins.top+t.margins.bottom+t.paddings.top+t.paddings.bottom+t.titleHeight+t.legendHeight}function r(t){return t.margins.left+t.margins.right+t.paddings.left+t.paddings.right}const o=["light-blue","blue","violet","red","orange","yellow","green","light-green","purple","magenta","light-grey","dark-grey"],l={bar:o,line:o,pie:o,percentage:o,heatmap:["#ebedf0","#c6e48b","#7bc96f","#239a3b","#196127"],donut:o},h=Math.PI/180;class d{constructor({parent:t=null,colors:e=[]}){this.parent=t,this.colors=e,this.titleName="",this.titleValue="",this.listValues=[],this.titleValueFirst=0,this.x=0,this.y=0,this.top=0,this.left=0,this.setup()}setup(){this.makeTooltip()}refresh(){this.fill(),this.calcPosition()}makeTooltip(){this.container=t.create("div",{inside:this.parent,className:"graph-svg-tip comparison",innerHTML:'<span class="title"></span>\n\t\t\t\t<ul class="data-point-list"></ul>\n\t\t\t\t<div class="svg-pointer"></div>'}),this.hideTip(),this.title=this.container.querySelector(".title"),this.dataPointList=this.container.querySelector(".data-point-list"),this.parent.addEventListener("mouseleave",(()=>{this.hideTip()}))}fill(){let e;this.index&&this.container.setAttribute("data-point-index",this.index),e=this.titleValueFirst?`<strong>${this.titleValue}</strong>${this.titleName}`:`${this.titleName}<strong>${this.titleValue}</strong>`,this.title.innerHTML=e,this.dataPointList.innerHTML="",this.listValues.map(((e,s)=>{const i=this.colors[s]||"black";let a=0===e.formatted||e.formatted?e.formatted:e.value,n=t.create("li",{styles:{"border-top":`3px solid ${i}`},innerHTML:`<strong style="display: block;">${0===a||a?a:""}</strong>\n\t\t\t\t\t${e.title?e.title:""}`});this.dataPointList.appendChild(n)}))}calcPosition(){let t=this.container.offsetWidth;this.top=this.y-this.container.offsetHeight-5,this.left=this.x-t/2;let e=this.parent.offsetWidth-t,s=this.container.querySelector(".svg-pointer");if(this.left<0)s.style.left=`calc(50% - ${-1*this.left}px)`,this.left=0;else if(this.left>e){let t=`calc(50% + ${this.left-e}px)`;s.style.left=t,this.left=e}else s.style.left="50%"}setValues(t,e,s={},i=[],a=-1){this.titleName=s.name,this.titleValue=s.value,this.listValues=i,this.x=t,this.y=e,this.titleValueFirst=s.valueFirst||0,this.index=a,this.refresh()}hideTip(){this.container.style.top="0px",this.container.style.left="0px",this.container.style.opacity="0"}showTip(){this.container.style.top=this.top+"px",this.container.style.left=this.left+"px",this.container.style.opacity="1"}}function c(t){return parseFloat(t.toFixed(2))}function p(t,e,s,i=!1){s||(s=i?t[0]:t[t.length-1]);let a=new Array(Math.abs(e)).fill(s);return t=i?a.concat(t):t.concat(a)}function u(t,e){return(t+"").length*e}function m(t,e){return{x:Math.sin(t*h)*e,y:Math.cos(t*h)*e}}function g(t,e=!1){return!Number.isNaN(t)&&(void 0!==t&&(!!Number.isFinite(t)&&!(e&&t<0)))}function f(t){let e,s,i;if(t instanceof Date)return new Date(t.getTime());if("object"!=typeof t||null===t)return t;for(i in e=Array.isArray(t)?[]:{},t)s=t[i],e[i]=f(s);return e}function y(t,e){let s,i;return t<=e?(s=e-t,i=t):(s=t-e,i=e),[s,i]}function b(t,e,s=e.length-t.length){return s>0?t=p(t,s):e=p(e,s),[t,e]}function x(t,e){if(t)return t.length>e?t.slice(0,e-3)+"...":t}function v(t,e){let s=[];for(let i=0;i<t.length;i++)s.push([t[i],e[i]]);let i=(t,e,s,i)=>{let a=((t,e)=>{let s=e[0]-t[0],i=e[1]-t[1];return{length:Math.sqrt(Math.pow(s,2)+Math.pow(i,2)),angle:Math.atan2(i,s)}})(e||t,s||t),n=a.angle+(i?Math.PI:0),r=.2*a.length;return[t[0]+Math.cos(n)*r,t[1]+Math.sin(n)*r]};return((t,e)=>t.reduce(((t,s,i,a)=>0===i?`${s[0]},${s[1]}`:`${t} ${e(s,i,a)}`),""))(s,((t,e,s)=>{let a=i(s[e-1],s[e-2],t),n=i(t,s[e-1],s[e+1],!0);return`C ${a[0]},${a[1]} ${n[0]},${n[1]} ${t[0]},${t[1]}`}))}const w={"light-blue":"#7cd6fd",blue:"#5e64ff",violet:"#743ee2",red:"#ff5858",orange:"#ffa00a",yellow:"#feef72",green:"#28a745","light-green":"#98d85b",purple:"#b554ff",magenta:"#ffa3ef",black:"#36114C",grey:"#bdd3e6","light-grey":"#f0f4f7","dark-grey":"#b8c2cc"};function A(t){return t>255?255:t<0?0:t}function k(t,e){let s=L(t),i=!1;"#"==s[0]&&(s=s.slice(1),i=!0);let a=parseInt(s,16),n=A((a>>16)+e),r=A((a>>8&255)+e);return(i?"#":"")+(A((255&a)+e)|r<<8|n<<16).toString(16)}const L=t=>/rgb[a]{0,1}\([\d, ]+\)/gim.test(t)?/\D+(\d*)\D+(\d*)\D+(\d*)/gim.exec(t).map(((t,e)=>0!==e?Number(t).toString(16):"#")).reduce(((t,e)=>`${t}${e}`)):w[t]||t;function T(t,e){return"string"==typeof t?(e||document).querySelector(t):t||null}function M(t,e){var s=document.createElementNS("http://www.w3.org/2000/svg",t);for(var i in e){var a=e[i];if("inside"===i)T(a).appendChild(s);else if("around"===i){var n=T(a);n.parentNode.insertBefore(s,n),s.appendChild(n)}else"styles"===i?"object"==typeof a&&Object.keys(a).map((t=>{s.style[t]=a[t]})):("className"===i&&(i="class"),"innerHTML"===i?s.textContent=a:s.setAttribute(i,a))}return s}function D(t,e,s,i){return M("stop",{inside:t,style:`stop-color: ${s}`,offset:e,"stop-opacity":i})}function P(t,e="",s){let i={className:t,transform:e};return s&&(i.inside=s),M("g",i)}function C(t,e="",s="none",i="none",a=2){return M("path",{className:e,d:t,styles:{stroke:s,fill:i,"stroke-width":a}})}function $(t,e,s=!1){let i="path-fill-gradient-"+e+"-"+(s?"lighter":"default"),a=function(t,e){return M("linearGradient",{inside:t,id:e,x1:0,x2:0,y1:0,y2:1})}(t,i),n=[1,.6,.2];return s&&(n=[.4,.2,0]),D(a,"0%",e,n[0]),D(a,"50%",e,n[1]),D(a,"100%",e,n[2]),i}function N(t,e,s,i,a,n="none",r={}){let o={className:t,x:e,y:s,width:i,height:i,rx:a,fill:n};return Object.keys(r).map((t=>{o[t]=r[t]})),M("rect",o)}function E(t,e,s,i,a={}){let n=a.fontSize||10;return M("text",{className:t,x:e,y:s,dy:(void 0!==a.dy?a.dy:n/2)+"px","font-size":n+"px",fill:a.fill||"#555b51","text-anchor":a.textAnchor||"start",innerHTML:i})}function O(t,e,s,i,a={}){a.stroke||(a.stroke="#dadada"),a.lineType||(a.lineType=""),a.shortenNumbers&&(e=function(t){let e;if("number"==typeof t)e=t;else if("string"==typeof t&&(e=Number(t),Number.isNaN(e)))return t;let s=Math.floor(Math.log10(Math.abs(e)));if(s<=2)return e;let i=Math.floor(s/3),a=Math.pow(10,s-3*i)*+(e/Math.pow(10,s)).toFixed(1);return Math.round(100*a)/100+" "+["","K","M","B","T"][i]}(e));let n=M("line",{className:"line-horizontal "+a.className+("dashed"===a.lineType?"dashed":""),x1:s,x2:i,y1:0,y2:0,styles:{stroke:a.stroke}}),r=M("text",{x:s<i?s-4:s+4,y:0,dy:"3px","font-size":"10px","text-anchor":s<i?"end":"start",innerHTML:e+""}),o=M("g",{transform:`translate(0, ${t})`,"stroke-opacity":1});return 0!==r&&"0"!==r||(o.style.stroke="rgba(27, 31, 35, 0.6)"),o.appendChild(n),o.appendChild(r),o}function S(t,e,s,i={}){g(t)||(t=0),i.pos||(i.pos="bottom"),i.offset||(i.offset=0),i.mode||(i.mode="span"),i.stroke||(i.stroke="#dadada"),i.className||(i.className="");let a=s+6,n="span"===i.mode?-6:s;return"tick"===i.mode&&"top"===i.pos&&(a=-6,n=0),function(t,e,s,i,a={}){a.stroke||(a.stroke="#dadada");let n=M("line",{className:"line-vertical "+a.className,x1:0,x2:0,y1:s,y2:i,styles:{stroke:a.stroke}}),r=M("text",{x:0,y:s>i?s+4:s-4-10,dy:"10px","font-size":"10px","text-anchor":"middle",innerHTML:e+""}),o=M("g",{transform:`translate(${t}, 0)`});return o.appendChild(n),o.appendChild(r),o}(t,e,a,n,{stroke:i.stroke,className:i.className,lineType:i.lineType})}let z={bar:t=>{let e;"rect"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode();return s.style.fill="#000000",s.style.opacity="0.4",e&&s.setAttribute("transform",e),s},dot:t=>{let e;"circle"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode(),i=t.getAttribute("r"),a=t.getAttribute("fill");return s.setAttribute("r",parseInt(i)+4),s.setAttribute("fill",a),s.style.opacity="0.6",e&&s.setAttribute("transform",e),s},heat_square:t=>{let e;"circle"!==t.nodeName&&(e=t.getAttribute("transform"),t=t.childNodes[0]);let s=t.cloneNode(),i=t.getAttribute("r"),a=t.getAttribute("fill");return s.setAttribute("r",parseInt(i)+4),s.setAttribute("fill",a),s.style.opacity="0.6",e&&s.setAttribute("transform",e),s}},W={bar:(t,e)=>{let s;"rect"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["x","y","width","height"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)},dot:(t,e)=>{let s;"circle"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["cx","cy"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)},heat_square:(t,e)=>{let s;"circle"!==t.nodeName&&(s=t.getAttribute("transform"),t=t.childNodes[0]);let i=["cx","cy"];Object.values(t.attributes).filter((t=>i.includes(t.name)&&t.specified)).map((t=>{e.setAttribute(t.name,t.nodeValue)})),s&&e.setAttribute("transform",s)}};function H(t,e,s,i){let a="string"==typeof e?e:e.join(", ");return[t,{transform:s.join(", ")},i,"easein","translate",{transform:a}]}function F(t,e,s){return H(t,[0,s],[0,e],350)}function R(t,e){return[t,{d:e},350,"easein"]}const I={ease:"0.25 0.1 0.25 1",linear:"0 0 1 1",easein:"0.1 0.8 0.2 1",easeout:"0 0 0.58 1",easeinout:"0.42 0 0.58 1"};function Y(t,e){t.style.transform=e,t.style.webkitTransform=e,t.style.msTransform=e,t.style.mozTransform=e,t.style.oTransform=e}function j(t,e){let s=[],i=[];e.map((t=>{let e,a,n=t[0],r=n.parentNode;t[0]=n,[e,a]=function(t,e,s,i="linear",a,n={}){let r=t.cloneNode(!0),o=t.cloneNode(!0);for(var l in e){let d;d="transform"===l?document.createElementNS("http://www.w3.org/2000/svg","animateTransform"):document.createElementNS("http://www.w3.org/2000/svg","animate");let c=n[l]||t.getAttribute(l),p=e[l],u={attributeName:l,from:c,to:p,begin:"0s",dur:s/1e3+"s",values:c+";"+p,keySplines:I[i],keyTimes:"0;1",calcMode:"spline",fill:"freeze"};for(var h in a&&(u.type=a),u)d.setAttribute(h,u[h]);r.appendChild(d),a?o.setAttribute(l,`translate(${p})`):o.setAttribute(l,p)}return[r,o]}(...t),s.push(a),i.push([e,r]),r&&r.replaceChild(e,n)}));let a=t.cloneNode(!0);return i.map(((t,i)=>{t[1]&&(t[1].replaceChild(s[i],t[0]),e[i][0]=s[i])})),a}class B{constructor(t,e){if(e=f(e),this.parent="string"==typeof t?document.querySelector(t):t,!(this.parent instanceof HTMLElement))throw new Error("No `parent` element to render on was provided.");this.rawChartArgs=e,this.title=e.title||"",this.type=e.type||"",this.realData=this.prepareData(e.data),this.data=this.prepareFirstData(this.realData),this.colors=this.validateColors(e.colors,this.type),this.config={showTooltip:1,showLegend:1,isNavigable:e.isNavigable||0,animate:void 0!==e.animate?e.animate:1,truncateLegends:e.truncateLegends||1},this.measures=JSON.parse(JSON.stringify(s));let i=this.measures;this.setMeasures(e),this.title.length||(i.titleHeight=0),this.config.showLegend||(i.legendHeight=0),this.argHeight=e.height||i.baseHeight,this.state={},this.options={},this.initTimeout=700,this.config.isNavigable&&(this.overlays=[]),this.configure(e)}prepareData(t){return t}prepareFirstData(t){return t}validateColors(t,e){const s=[];return(t=(t||[]).concat(l[e])).forEach((t=>{const e=L(t);!function(t){return/(^\s*)(#)((?:[A-Fa-f0-9]{3}){1,2})$/i.test(t)||/(^\s*)(rgb|hsl)(a?)[(]\s*([\d.]+\s*%?)\s*,\s*([\d.]+\s*%?)\s*,\s*([\d.]+\s*%?)\s*(?:,\s*([\d.]+)\s*)?[)]$/i.test(t)}(e)?console.warn('"'+t+'" is not a valid color.'):s.push(e)})),s}setMeasures(){}configure(){let t=this.argHeight;this.baseHeight=t,this.height=t-n(this.measures),this.boundDrawFn=()=>this.draw(!0),ResizeObserver&&(this.resizeObserver=new ResizeObserver(this.boundDrawFn),this.resizeObserver.observe(this.parent)),window.addEventListener("resize",this.boundDrawFn),window.addEventListener("orientationchange",this.boundDrawFn)}destroy(){this.resizeObserver&&this.resizeObserver.disconnect(),window.removeEventListener("resize",this.boundDrawFn),window.removeEventListener("orientationchange",this.boundDrawFn)}setup(){this.makeContainer(),this.updateWidth(),this.makeTooltip(),this.draw(!1,!0)}makeContainer(){this.parent.innerHTML="";let e={inside:this.parent,className:"chart-container"};this.independentWidth&&(e.styles={width:this.independentWidth+"px"}),this.container=t.create("div",e)}makeTooltip(){this.tip=new d({parent:this.container,colors:this.colors}),this.bindTooltip()}bindTooltip(){}draw(t=!1,e=!1){t&&null===this.parent.offsetParent||(this.updateWidth(),this.calc(t),this.makeChartArea(),this.setupComponents(),this.components.forEach((t=>t.setup(this.drawArea))),this.render(this.components,!1),e&&(this.data=this.realData,setTimeout((()=>{this.update(this.data)}),this.initTimeout)),this.renderLegend(),this.setupNavigation(e))}calc(){}updateWidth(){var t,e,s;this.baseWidth=(t=this.parent,e=window.getComputedStyle(t),s=parseFloat(e.paddingLeft)+parseFloat(e.paddingRight),t.clientWidth-s),this.width=this.baseWidth-r(this.measures)}makeChartArea(){this.svg&&this.container.removeChild(this.svg);let t=this.measures;var e,s,n,r;this.svg=(e=this.container,s="frappe-chart chart",n=this.baseWidth,r=this.baseHeight,M("svg",{className:s,inside:e,width:n,height:r})),this.svgDefs=M("defs",{inside:this.svg}),this.title.length&&(this.titleEL=E("title",t.margins.left,t.margins.top,this.title,{fontSize:t.titleFontSize,fill:"#666666",dy:t.titleFontSize}));let o=i(t);this.drawArea=P(this.type+"-chart chart-draw-area",`translate(${a(t)}, ${o})`),this.config.showLegend&&(o+=this.height+t.paddings.bottom,this.legendArea=P("chart-legend",`translate(${a(t)}, ${o})`)),this.title.length&&this.svg.appendChild(this.titleEL),this.svg.appendChild(this.drawArea),this.config.showLegend&&this.svg.appendChild(this.legendArea),this.updateTipOffset(a(t),i(t))}updateTipOffset(t,e){this.tip.offset={x:t,y:e}}setupComponents(){this.components=new Map}update(t){t||console.error("No data to update."),this.data=this.prepareData(t),this.calc(),this.render(this.components,this.config.animate),this.renderLegend()}render(t=this.components,e=!0){this.config.isNavigable&&this.overlays.map((t=>t.parentNode.removeChild(t)));let s=[];t.forEach((t=>{s=s.concat(t.update(e))})),s.length>0?(!function(t,e,s){if(0===s.length)return;let i=j(e,s);e.parentNode==t&&(t.removeChild(e),t.appendChild(i)),setTimeout((()=>{i.parentNode==t&&(t.removeChild(i),t.appendChild(e))}),250)}(this.container,this.svg,s),setTimeout((()=>{t.forEach((t=>t.make())),this.updateNav()}),400)):(t.forEach((t=>t.make())),this.updateNav())}updateNav(){this.config.isNavigable&&(this.makeOverlay(),this.bindUnits())}renderLegend(){}setupNavigation(t=!1){this.config.isNavigable&&t&&(this.bindOverlay(),this.keyActions={13:this.onEnterKey.bind(this),37:this.onLeftArrow.bind(this),38:this.onUpArrow.bind(this),39:this.onRightArrow.bind(this),40:this.onDownArrow.bind(this)},document.addEventListener("keydown",(t=>{var e,s;e=this.container,(s=e.getBoundingClientRect()).top>=0&&s.left>=0&&s.bottom<=(window.innerHeight||document.documentElement.clientHeight)&&s.right<=(window.innerWidth||document.documentElement.clientWidth)&&(t=t||window.event,this.keyActions[t.keyCode]&&this.keyActions[t.keyCode]())})))}makeOverlay(){}updateOverlay(){}bindOverlay(){}bindUnits(){}onLeftArrow(){}onRightArrow(){}onUpArrow(){}onDownArrow(){}onEnterKey(){}addDataPoint(){}removeDataPoint(){}getDataPoint(){}setCurrentDataPoint(){}updateDataset(){}export(){let e=function(e){let s=e.cloneNode(!0);s.classList.add("chart-container"),s.setAttribute("xmlns","http://www.w3.org/2000/svg"),s.setAttribute("xmlns:xlink","http://www.w3.org/1999/xlink");let i=t.create("style",{innerHTML:".chart-container{position:relative;font-family:-apple-system,BlinkMacSystemFont,'Segoe UI','Roboto','Oxygen','Ubuntu','Cantarell','Fira Sans','Droid Sans','Helvetica Neue',sans-serif}.chart-container .axis,.chart-container .chart-label{fill:#555b51}.chart-container .axis line,.chart-container .chart-label line{stroke:#dadada}.chart-container .dataset-units circle{stroke:#fff;stroke-width:2}.chart-container .dataset-units path{fill:none;stroke-opacity:1;stroke-width:2px}.chart-container .dataset-path{stroke-width:2px}.chart-container .path-group path{fill:none;stroke-opacity:1;stroke-width:2px}.chart-container line.dashed{stroke-dasharray:5,3}.chart-container .axis-line .specific-value{text-anchor:start}.chart-container .axis-line .y-line{text-anchor:end}.chart-container .axis-line .x-line{text-anchor:middle}.chart-container .legend-dataset-text{fill:#6c7680;font-weight:600}.graph-svg-tip{position:absolute;z-index:99999;padding:10px;font-size:12px;color:#959da5;text-align:center;background:rgba(0,0,0,.8);border-radius:3px}.graph-svg-tip ul{padding-left:0;display:flex}.graph-svg-tip ol{padding-left:0;display:flex}.graph-svg-tip ul.data-point-list li{min-width:90px;flex:1;font-weight:600}.graph-svg-tip strong{color:#dfe2e5;font-weight:600}.graph-svg-tip .svg-pointer{position:absolute;height:5px;margin:0 0 0 -5px;content:' ';border:5px solid transparent;border-top-color:rgba(0,0,0,.8)}.graph-svg-tip.comparison{padding:0;text-align:left;pointer-events:none}.graph-svg-tip.comparison .title{display:block;padding:10px;margin:0;font-weight:600;line-height:1;pointer-events:none}.graph-svg-tip.comparison ul{margin:0;white-space:nowrap;list-style:none}.graph-svg-tip.comparison li{display:inline-block;padding:5px 10px}"});s.insertBefore(i,s.firstChild);let a=t.create("div");return a.appendChild(s),a.innerHTML}(this.svg);!function(t,e){var s=document.createElement("a");s.style="display: none";var i=new Blob(e,{type:"image/svg+xml; charset=utf-8"}),a=window.URL.createObjectURL(i);s.href=a,s.download=t,document.body.appendChild(s),s.click(),setTimeout((function(){document.body.removeChild(s),window.URL.revokeObjectURL(a)}),300)}(this.title||"Chart",[e])}}class V extends B{constructor(t,e){super(t,e)}configure(t){super.configure(t),this.config.formatTooltipY=(t.tooltipOptions||{}).formatTooltipY,this.config.maxSlices=t.maxSlices||20,this.config.maxLegendPoints=t.maxLegendPoints||20}calc(){let t=this.state,e=this.config.maxSlices;t.sliceTotals=[];let s=this.data.labels.map(((t,e)=>{let s=0;return this.data.datasets.map((t=>{s+=t.values[e]})),[s,t]})).filter((t=>t[0]>=0)),i=s;if(s.length>e){s.sort(((t,e)=>e[0]-t[0])),i=s.slice(0,e-1);let t=s.slice(e-1),a=0;t.map((t=>{a+=t[0]})),i.push([a,"Rest"]),this.colors[e-1]="grey"}t.labels=[],i.map((e=>{t.sliceTotals.push(function(t){return Number(Math.round(t+"e4")+"e-4")}(e[0])),t.labels.push(e[1])})),t.grandTotal=t.sliceTotals.reduce(((t,e)=>t+e),0),this.center={x:this.width/2,y:this.height/2}}renderLegend(){let t=this.state;this.legendArea.textContent="",this.legendTotals=t.sliceTotals.slice(0,this.config.maxLegendPoints);let e=0,s=0;this.legendTotals.map(((i,a)=>{let n=150,o=Math.floor((this.width-r(this.measures))/n);this.legendTotals.length<o&&(n=this.width/this.legendTotals.length),e>o&&(e=0,s+=20);let l=n*e+5,h=this.config.truncateLegends?x(t.labels[a],n/10):t.labels[a],d=this.config.formatTooltipY?this.config.formatTooltipY(i):i,c=function(t,e,s,i="none",a,n=!1){let r={className:"legend-dot",cx:0,cy:0,r:s,fill:i},o=M("text",{className:"legend-dataset-text",x:0,y:0,dx:"10px",dy:10/3+"px","font-size":"12px","text-anchor":"start",fill:"#555b51",innerHTML:a=n?x(a,15):a}),l=M("g",{transform:`translate(${t}, ${e})`});return l.appendChild(M("circle",r)),l.appendChild(o),l}(l,s,5,this.colors[a],`${h}: ${d}`,!1);this.legendArea.appendChild(c),e++}))}}const U=["January","February","March","April","May","June","July","August","September","October","November","December"],G=["Sun","Mon","Tue","Wed","Thu","Fri","Sat"];function q(t){let e=new Date(t);return e.setMinutes(e.getMinutes()-e.getTimezoneOffset()),e}function X(t){let e=t.getDate(),s=t.getMonth()+1;return[t.getFullYear(),(s>9?"":"0")+s,(e>9?"":"0")+e].join("-")}function _(t){return new Date(t.getTime())}function J(t,e){let s=tt(t);return Math.ceil(function(t,e){let s=864e5;return(q(e)-q(t))/s}(s,e)/7)}function K(t,e){return t.getMonth()===e.getMonth()&&t.getFullYear()===e.getFullYear()}function Q(t,e=!1){let s=U[t];return e?s.slice(0,3):s}function Z(t,e){return new Date(e,t+1,0)}function tt(t){let e=_(t);const s=e.getDay();return 0!==s&&et(e,-1*s),e}function et(t,e){t.setDate(t.getDate()+e)}class st{constructor({layerClass:t="",layerTransform:e="",constants:s,getData:i,makeElements:a,animateElements:n}){this.layerTransform=e,this.constants=s,this.makeElements=a,this.getData=i,this.animateElements=n,this.store=[],this.labels=[],this.layerClass=t,this.layerClass="function"==typeof this.layerClass?this.layerClass():this.layerClass,this.refresh()}refresh(t){this.data=t||this.getData()}setup(t){this.layer=P(this.layerClass,this.layerTransform,t)}make(){this.render(this.data),this.oldData=this.data}render(t){this.store=this.makeElements(t),this.layer.textContent="",this.store.forEach((t=>{this.layer.appendChild(t)})),this.labels.forEach((t=>{this.layer.appendChild(t)}))}update(t=!0){this.refresh();let e=[];return t&&(e=this.animateElements(this.data)||[]),e}}let it={donutSlices:{layerClass:"donut-slices",makeElements:t=>t.sliceStrings.map(((e,s)=>{let i=C(e,"donut-path",t.colors[s],"none",t.strokeWidth);return i.style.transition="transform .3s;",i})),animateElements(t){return this.store.map(((e,s)=>R(e,t.sliceStrings[s])))}},pieSlices:{layerClass:"pie-slices",makeElements:t=>t.sliceStrings.map(((e,s)=>{let i=C(e,"pie-path","none",t.colors[s]);return i.style.transition="transform .3s;",i})),animateElements(t){return this.store.map(((e,s)=>R(e,t.sliceStrings[s])))}},percentageBars:{layerClass:"percentage-bars",makeElements(t){return t.xPositions.map(((e,s)=>function(t,e,s,i,a=2,n="none"){return M("rect",{className:"percentage-bar",x:t,y:e,width:s,height:i,fill:n,styles:{stroke:k(n,-25),"stroke-dasharray":`0, ${i+s}, ${s}, ${i}`,"stroke-width":a}})}(e,0,t.widths[s],this.constants.barHeight,this.constants.barDepth,t.colors[s])))},animateElements(t){if(t)return[]}},yAxis:{layerClass:"y axis",makeElements(t){return t.positions.map(((e,s)=>function(t,e,s,i={}){g(t)||(t=0),i.pos||(i.pos="left"),i.offset||(i.offset=0),i.mode||(i.mode="span"),i.stroke||(i.stroke="#dadada"),i.className||(i.className="");let a=-6,n="span"===i.mode?s+6:0;return"tick"===i.mode&&"right"===i.pos&&(a=s+6,n=s),a+=i.offset,n+=i.offset,O(t,e,a,n,{stroke:i.stroke,className:i.className,lineType:i.lineType,shortenNumbers:i.shortenNumbers})}(e,t.labels[s],this.constants.width,{mode:this.constants.mode,pos:this.constants.pos,shortenNumbers:this.constants.shortenNumbers})))},animateElements(t){let e=t.positions,s=t.labels,i=this.oldData.positions,a=this.oldData.labels;return[i,e]=b(i,e),[a,s]=b(a,s),this.render({positions:i,labels:s}),this.store.map(((t,s)=>F(t,e[s],i[s])))}},xAxis:{layerClass:"x axis",makeElements(t){return t.positions.map(((e,s)=>S(e,t.calcLabels[s],this.constants.height,{mode:this.constants.mode,pos:this.constants.pos})))},animateElements(t){let e=t.positions,s=t.calcLabels,i=this.oldData.positions,a=this.oldData.calcLabels;return[i,e]=b(i,e),[a,s]=b(a,s),this.render({positions:i,calcLabels:s}),this.store.map(((t,s)=>function(t,e,s){return H(t,[s,0],[e,0],350)}(t,e[s],i[s])))}},yMarkers:{layerClass:"y-markers",makeElements(t){return t.map((t=>function(t,e,s,i={}){i.labelPos||(i.labelPos="right");let a=M("text",{className:"chart-label",x:"left"===i.labelPos?4:s-u(e,5)-4,y:0,dy:"-5px","font-size":"10px","text-anchor":"start",innerHTML:e+""}),n=O(t,"",0,s,{stroke:i.stroke||"#dadada",className:i.className||"",lineType:i.lineType});return n.appendChild(a),n}(t.position,t.label,this.constants.width,{labelPos:t.options.labelPos,mode:"span",lineType:"dashed"})))},animateElements(t){[this.oldData,t]=b(this.oldData,t);let e=t.map((t=>t.position)),s=t.map((t=>t.label)),i=t.map((t=>t.options)),a=this.oldData.map((t=>t.position));return this.render(a.map(((t,e)=>({position:a[e],label:s[e],options:i[e]})))),this.store.map(((t,s)=>F(t,e[s],a[s])))}},yRegions:{layerClass:"y-regions",makeElements(t){return t.map((t=>function(t,e,s,i,a={}){let n=t-e,r=M("rect",{className:"bar mini",styles:{fill:"rgba(228, 234, 239, 0.49)",stroke:"#dadada","stroke-dasharray":`${s}, ${n}`},x:0,y:0,width:s,height:n});a.labelPos||(a.labelPos="right");let o=M("text",{className:"chart-label",x:"left"===a.labelPos?4:s-u(i+"",4.5)-4,y:0,dy:"-5px","font-size":"10px","text-anchor":"start",innerHTML:i+""}),l=M("g",{transform:`translate(0, ${e})`});return l.appendChild(r),l.appendChild(o),l}(t.startPos,t.endPos,this.constants.width,t.label,{labelPos:t.options.labelPos})))},animateElements(t){[this.oldData,t]=b(this.oldData,t);let e=t.map((t=>t.endPos)),s=t.map((t=>t.label)),i=t.map((t=>t.startPos)),a=t.map((t=>t.options)),n=this.oldData.map((t=>t.endPos)),r=this.oldData.map((t=>t.startPos));this.render(n.map(((t,e)=>({startPos:r[e],endPos:n[e],label:s[e],options:a[e]}))));let o=[];return this.store.map(((t,s)=>{o=o.concat(function(t,e,s,i){let a=e-s,n=t.childNodes[0],r=n.getAttribute("width");return[[n,{height:a,"stroke-dasharray":`${r}, ${a}`},350,"easein"],H(t,[0,i],[0,s],350)]}(t,i[s],e[s],n[s]))})),o}},heatDomain:{layerClass:function(){return"heat-domain domain-"+this.constants.index},makeElements(t){let{index:e,colWidth:s,rowHeight:i,squareSize:a,radius:n,xTranslate:r}=this.constants,o=r,l=0;return this.serializedSubDomains=[],t.cols.map(((t,r)=>{1===r&&this.labels.push(E("domain-name",o,-12,Q(e,!0).toUpperCase(),{fontSize:9})),t.map(((t,e)=>{if(t.fill){let s={"data-date":t.yyyyMmDd,"data-value":t.dataValue,"data-day":e},i=N("day",o,l,a,n,t.fill,s);this.serializedSubDomains.push(i)}l+=i})),l=0,o+=s})),this.serializedSubDomains},animateElements(t){if(t)return[]}},barGraph:{layerClass:function(){return"dataset-units dataset-bars dataset-"+this.constants.index},makeElements(t){let e=this.constants;return this.unitType="bar",this.units=t.yPositions.map(((s,i)=>function(t,e,s,i,a="",n=0,r=0,o={}){let[l,h]=y(e,o.zeroLine);h-=r,0===l&&(l=o.minHeight,h-=o.minHeight),g(t)||(t=0),g(h)||(h=0),g(l,!0)||(l=0),g(s,!0)||(s=0);let d=M("rect",{className:"bar mini",style:`fill: ${i}`,"data-point-index":n,x:t,y:h,width:s,height:l});if((a+="")||a.length){d.setAttribute("y",0),d.setAttribute("x",0);let e=M("text",{className:"data-point-value",x:s/2,y:0,dy:"-5px","font-size":"10px","text-anchor":"middle",innerHTML:a}),i=M("g",{"data-point-index":n,transform:`translate(${t}, ${h})`});return i.appendChild(d),i.appendChild(e),i}return d}(t.xPositions[i],s,t.barWidth,e.color,t.labels[i],i,t.offsets[i],{zeroLine:t.zeroLine,barsWidth:t.barsWidth,minHeight:e.minHeight}))),this.units},animateElements(t){let e=t.xPositions,s=t.yPositions,i=t.offsets,a=t.labels,n=this.oldData.xPositions,r=this.oldData.yPositions,o=this.oldData.offsets,l=this.oldData.labels;[n,e]=b(n,e),[r,s]=b(r,s),[o,i]=b(o,i),[l,a]=b(l,a),this.render({xPositions:n,yPositions:r,offsets:o,labels:a,zeroLine:this.oldData.zeroLine,barsWidth:this.oldData.barsWidth,barWidth:this.oldData.barWidth});let h=[];return this.store.map(((a,n)=>{h=h.concat(function(t,e,s,i,a=0,n={}){let[r,o]=y(s,n.zeroLine);if(o-=a,"rect"!==t.nodeName){let s=[t.childNodes[0],{width:i,height:r},350,"easein"],a=t.getAttribute("transform").split("(")[1].slice(0,-1);return[s,H(t,a,[e,o],350)]}return[[t,{width:i,height:r,x:e,y:o},350,"easein"]]}(a,e[n],s[n],t.barWidth,i[n],{zeroLine:t.zeroLine}))})),h}},lineGraph:{layerClass:function(){return"dataset-units dataset-line dataset-"+this.constants.index},makeElements(t){let e=this.constants;return this.unitType="dot",this.paths={},e.hideLine||(this.paths=function(t,e,s,i={},a={}){let n=e.map(((e,s)=>t[s]+","+e)).join("L");i.spline&&(n=v(t,e));let r=C("M"+n,"line-graph-path",s);if(i.heatline){let t=$(a.svgDefs,s);r.style.stroke=`url(#${t})`}let o={path:r};if(i.regionFill){let e=$(a.svgDefs,s,!0),i=`M${t[0]},${a.zeroLine}L`+n+`L${t.slice(-1)[0]},${a.zeroLine}`;o.region=C(i,"region-fill","none",`url(#${e})`)}return o}(t.xPositions,t.yPositions,e.color,{heatline:e.heatline,regionFill:e.regionFill,spline:e.spline},{svgDefs:e.svgDefs,zeroLine:t.zeroLine})),this.units=[],e.hideDots||(this.units=t.yPositions.map(((s,i)=>function(t,e,s,i,a="",n=0){let r=M("circle",{style:`fill: ${i}`,"data-point-index":n,cx:t,cy:e,r:s});if((a+="")||a.length){r.setAttribute("cy",0),r.setAttribute("cx",0);let i=M("text",{className:"data-point-value",x:0,y:0,dy:-5-s+"px","font-size":"10px","text-anchor":"middle",innerHTML:a}),o=M("g",{"data-point-index":n,transform:`translate(${t}, ${e})`});return o.appendChild(r),o.appendChild(i),o}return r}(t.xPositions[i],s,t.radius,e.color,e.valuesOverPoints?t.values[i]:"",i)))),Object.values(this.paths).concat(this.units)},animateElements(t){let e=t.xPositions,s=t.yPositions,i=t.values,a=this.oldData.xPositions,n=this.oldData.yPositions,r=this.oldData.values;[a,e]=b(a,e),[n,s]=b(n,s),[r,i]=b(r,i),this.render({xPositions:a,yPositions:n,values:i,zeroLine:this.oldData.zeroLine,radius:this.oldData.radius});let o=[];return Object.keys(this.paths).length&&(o=o.concat(function(t,e,s,i,a){let n=[],r=s.map(((t,s)=>e[s]+","+t)).join("L");a&&(r=v(e,s));const o=[t.path,{d:"M"+r},350,"easein"];if(n.push(o),t.region){let s=`${e[0]},${i}L`,a=`L${e.slice(-1)[0]}, ${i}`;const o=[t.region,{d:"M"+s+r+a},350,"easein"];n.push(o)}return n}(this.paths,e,s,t.zeroLine,this.constants.spline))),this.units.length&&this.units.map(((t,i)=>{o=o.concat(function(t,e,s){if("circle"!==t.nodeName){let i=t.getAttribute("transform").split("(")[1].slice(0,-1);return[H(t,i,[e,s],350)]}return[[t,{cx:e,cy:s},350,"easein"]]}(t,e[i],s[i]))})),o}}};function at(t,e,s){let i=Object.keys(it).filter((e=>t.includes(e))),a=it[i[0]];return Object.assign(a,{constants:e,getData:s}),new st(a)}function nt(t){if(0===t)return[0,0];if(isNaN(t))return{mantissa:-6755399441055744,exponent:972};var e=t>0?1:-1;if(!isFinite(t))return{mantissa:4503599627370496*e,exponent:972};t=Math.abs(t);var s=Math.floor(Math.log10(t));return[e*(t/Math.pow(10,s)),s]}function rt(t,e=0){let[s,i]=nt(t),a=e?e/Math.pow(10,i):0;s=s.toFixed(6);let n=function(t,e=0){let s=Math.ceil(t),i=Math.floor(e),a=s-i,n=a,r=1;a>5&&(a%2!=0&&(s++,a=s-i),n=a/2,r=2),a<=2&&(n=4,r=a/n),0===a&&(n=5,r=1);let o=[];for(var l=0;l<=n;l++)o.push(i+r*l);return o}(s,a);return n=n.map((t=>t*Math.pow(10,i))),n}function ot(t){return t[1]-t[0]}function lt(t,e){return c(e.zeroLine-t*e.scaleMultiplier)}class ht extends B{constructor(t,e){super(t,e),this.barOptions=e.barOptions||{},this.lineOptions=e.lineOptions||{},this.type=e.type||"line",this.init=1,this.setup()}setMeasures(){this.data.datasets.length<=1&&(this.config.showLegend=0,this.measures.paddings.bottom=30)}configure(t){super.configure(t),t.axisOptions=t.axisOptions||{},t.tooltipOptions=t.tooltipOptions||{},this.config.xAxisMode=t.axisOptions.xAxisMode||"span",this.config.yAxisMode=t.axisOptions.yAxisMode||"span",this.config.xIsSeries=t.axisOptions.xIsSeries||0,this.config.shortenYAxisNumbers=t.axisOptions.shortenYAxisNumbers||0,this.config.formatTooltipX=t.tooltipOptions.formatTooltipX,this.config.formatTooltipY=t.tooltipOptions.formatTooltipY,this.config.valuesOverPoints=t.valuesOverPoints}prepareData(t=this.data){return function(t,e){t.labels=t.labels||[];let s=t.labels.length,i=t.datasets,a=new Array(s).fill(0);return i||(i=[{values:a}]),i.map((t=>{if(t.values){let e=t.values;e=e.map((t=>isNaN(t)?0:t)),e=e.length>s?e.slice(0,s):p(e,s-e.length,0),t.values=e}else t.values=a;t.chartType||(t.chartType=e)})),t.yRegions&&t.yRegions.map((t=>{t.end<t.start&&([t.start,t.end]=[t.end,t.start])})),t}(t,this.type)}prepareFirstData(t=this.data){return function(t){let e=t.labels.length,s=new Array(e).fill(0),i={labels:t.labels.slice(0,-1),datasets:t.datasets.map((t=>({name:"",values:s.slice(0,-1),chartType:t.chartType})))};return t.yMarkers&&(i.yMarkers=[{value:0,label:""}]),t.yRegions&&(i.yRegions=[{start:0,end:0,label:""}]),i}(t)}calc(t=!1){this.calcXPositions(),t||this.calcYAxisParameters(this.getAllYValues(),"line"===this.type),this.makeDataByIndex()}calcXPositions(){let t=this.state,e=this.data.labels;t.datasetLength=e.length,t.unitWidth=this.width/t.datasetLength,t.xOffset=t.unitWidth/2,t.xAxis={labels:e,positions:e.map(((e,s)=>c(t.xOffset+s*t.unitWidth)))}}calcYAxisParameters(t,e="false"){const s=function(t,e=!1){let s=Math.max(...t),i=Math.min(...t),a=[];function n(t,e){let s=rt(t),i=s[1]-s[0],a=0;for(var n=1;a<e;n++)a+=i,s.unshift(-1*a);return s}if(s>=0&&i>=0)nt(s)[1],a=e?rt(s,i):rt(s);else if(s>0&&i<0){let t=Math.abs(i);s>=t?(nt(s)[1],a=n(s,t)):(nt(t)[1],a=n(t,s).reverse().map((t=>-1*t)))}else if(s<=0&&i<=0){let t=Math.abs(i),n=Math.abs(s);nt(t)[1],a=e?rt(t,n):rt(t),a=a.reverse().map((t=>-1*t))}return a}(t,e),i=this.height/((a=s)[a.length-1]-a[0]);var a;const n=ot(s)*i,r=this.height-function(t){let e,s=ot(t);e=t.indexOf(0)>=0?t.indexOf(0):t[0]>0?-1*t[0]/s:-1*t[t.length-1]/s+(t.length-1);return e}(s)*n;this.state.yAxis={labels:s,positions:s.map((t=>r-t*i)),scaleMultiplier:i,zeroLine:r},this.calcDatasetPoints(),this.calcYExtremes(),this.calcYRegions()}calcDatasetPoints(){let t=this.state,e=e=>e.map((e=>lt(e,t.yAxis)));t.datasets=this.data.datasets.map(((t,s)=>{let i=t.values,a=t.cumulativeYs||[];return{name:t.name&&t.name.replace(/<|>|&/g,(t=>"&"==t?"&amp;":"<"==t?"&lt;":"&gt;")),index:s,chartType:t.chartType,values:i,yPositions:e(i),cumulativeYs:a,cumulativeYPos:e(a)}}))}calcYExtremes(){let t=this.state;this.barOptions.stacked?t.yExtremes=t.datasets[t.datasets.length-1].cumulativeYPos:(t.yExtremes=new Array(t.datasetLength).fill(9999),t.datasets.map((e=>{e.yPositions.map(((e,s)=>{e<t.yExtremes[s]&&(t.yExtremes[s]=e)}))})))}calcYRegions(){let t=this.state;this.data.yMarkers&&(this.state.yMarkers=this.data.yMarkers.map((e=>(e.position=lt(e.value,t.yAxis),e.options||(e.options={}),e)))),this.data.yRegions&&(this.state.yRegions=this.data.yRegions.map((e=>(e.startPos=lt(e.start,t.yAxis),e.endPos=lt(e.end,t.yAxis),e.options||(e.options={}),e))))}getAllYValues(){let t="values";if(this.barOptions.stacked){t="cumulativeYs";let e=new Array(this.state.datasetLength).fill(0);this.data.datasets.map(((s,i)=>{let a=this.data.datasets[i].values;s[t]=e=e.map(((t,e)=>t+a[e]))}))}let e=this.data.datasets.map((e=>e[t]));return this.data.yMarkers&&e.push(this.data.yMarkers.map((t=>t.value))),this.data.yRegions&&this.data.yRegions.map((t=>{e.push([t.end,t.start])})),[].concat(...e)}setupComponents(){let t=[["yAxis",{mode:this.config.yAxisMode,width:this.width,shortenNumbers:this.config.shortenYAxisNumbers},function(){return this.state.yAxis}.bind(this)],["xAxis",{mode:this.config.xAxisMode,height:this.height},function(){let t=this.state;return t.xAxis.calcLabels=function(t,e=[],s=!0){let i=t/e.length;i<=0&&(i=1);let a,n=i/7;if(s){let t=Math.max(...e.map((t=>t.length)));a=Math.ceil(t/n)}return e.map(((t,e)=>((t+="").length>n&&(s?e%a!=0&&(t=""):t=n-3>0?t.slice(0,n-3)+" ...":t.slice(0,n)+".."),t)))}(this.width,t.xAxis.labels,this.config.xIsSeries),t.xAxis}.bind(this)],["yRegions",{width:this.width,pos:"right"},function(){return this.state.yRegions}.bind(this)]],e=this.state.datasets.filter((t=>"bar"===t.chartType)),s=this.state.datasets.filter((t=>"line"===t.chartType)),i=e.map((t=>{let s=t.index;return["barGraph-"+t.index,{index:s,color:this.colors[s],stacked:this.barOptions.stacked,valuesOverPoints:this.config.valuesOverPoints,minHeight:0*this.height},function(){let t=this.state,i=t.datasets[s],a=this.barOptions.stacked,n=this.barOptions.spaceRatio||.5,r=t.unitWidth*(1-n),o=r/(a?1:e.length),l=t.xAxis.positions.map((t=>t-r/2));a||(l=l.map((t=>t+o*s)));let h=new Array(t.datasetLength).fill("");this.config.valuesOverPoints&&(h=a&&i.index===t.datasets.length-1?i.cumulativeYs:i.values);let d=new Array(t.datasetLength).fill(0);return a&&(d=i.yPositions.map(((t,e)=>t-i.cumulativeYPos[e]))),{xPositions:l,yPositions:i.yPositions,offsets:d,labels:h,zeroLine:t.yAxis.zeroLine,barsWidth:r,barWidth:o}}.bind(this)]})),a=s.map((t=>{let e=t.index;return["lineGraph-"+t.index,{index:e,color:this.colors[e],svgDefs:this.svgDefs,heatline:this.lineOptions.heatline,regionFill:this.lineOptions.regionFill,spline:this.lineOptions.spline,hideDots:this.lineOptions.hideDots,hideLine:this.lineOptions.hideLine,valuesOverPoints:this.config.valuesOverPoints},function(){let t=this.state,s=t.datasets[e],i=t.yAxis.positions[0]<t.yAxis.zeroLine?t.yAxis.positions[0]:t.yAxis.zeroLine;return{xPositions:t.xAxis.positions,yPositions:s.yPositions,values:s.values,zeroLine:i,radius:this.lineOptions.dotSize||4}}.bind(this)]})),n=[["yMarkers",{width:this.width,pos:"right"},function(){return this.state.yMarkers}.bind(this)]];t=t.concat(i,a,n);let r=["yMarkers","yRegions"];this.dataUnitComponents=[],this.components=new Map(t.filter((t=>!r.includes(t[0])||this.state[t[0]])).map((t=>{let e=at(...t);return(t[0].includes("lineGraph")||t[0].includes("barGraph"))&&this.dataUnitComponents.push(e),[t[0],e]})))}makeDataByIndex(){this.dataByIndex={};let t=this.state,e=this.config.formatTooltipX,s=this.config.formatTooltipY;t.xAxis.labels.map(((i,a)=>{let n=this.state.datasets.map(((t,e)=>{let i=t.values[a];return{title:t.name,value:i,yPos:t.yPositions[a],color:this.colors[e],formatted:s?s(i):i}}));this.dataByIndex[a]={label:i,formattedLabel:e?e(i):i,xPos:t.xAxis.positions[a],values:n,yExtreme:t.yExtremes[a]}}))}bindTooltip(){this.container.addEventListener("mousemove",(t=>{let s=this.measures,n=e(this.container),r=t.pageX-n.left-a(s),o=t.pageY-n.top;o<this.height+i(s)&&o>i(s)?this.mapTooltipXPosition(r):this.tip.hideTip()}))}mapTooltipXPosition(t){let e=this.state;if(!e.yExtremes)return;let s=function(t,e,s=!1){let i=e.reduce((function(e,s){return Math.abs(s-t)<Math.abs(e-t)?s:e}),[]);return s?e.indexOf(i):i}(t,e.xAxis.positions,!0);if(s>=0){let t=this.dataByIndex[s];this.tip.setValues(t.xPos+this.tip.offset.x,t.yExtreme+this.tip.offset.y,{name:t.formattedLabel,value:""},t.values,s),this.tip.showTip()}}renderLegend(){let t=this.data;t.datasets.length>1&&(this.legendArea.textContent="",t.datasets.map(((t,e)=>{let s=function(t,e,s,i="none",a,n=!1){let r={className:"legend-bar",x:0,y:0,width:s,height:"2px",fill:i},o=M("text",{className:"legend-dataset-text",x:0,y:0,dy:"20px","font-size":"12px","text-anchor":"start",fill:"#555b51",innerHTML:a=n?x(a,15):a}),l=M("g",{transform:`translate(${t}, ${e})`});return l.appendChild(M("rect",r)),l.appendChild(o),l}(100*e,"0",100,this.colors[e],t.name,this.config.truncateLegends);this.legendArea.appendChild(s)})))}makeOverlay(){this.init?this.init=0:(this.overlayGuides&&this.overlayGuides.forEach((t=>{let e=t.overlay;e.parentNode.removeChild(e)})),this.overlayGuides=this.dataUnitComponents.map((t=>({type:t.unitType,overlay:void 0,units:t.units}))),void 0===this.state.currentIndex&&(this.state.currentIndex=this.state.datasetLength-1),this.overlayGuides.map((t=>{let e=t.units[this.state.currentIndex];t.overlay=z[t.type](e),this.drawArea.appendChild(t.overlay)})))}updateOverlayGuides(){this.overlayGuides&&this.overlayGuides.forEach((t=>{let e=t.overlay;e.parentNode.removeChild(e)}))}bindOverlay(){this.parent.addEventListener("data-select",(()=>{this.updateOverlay()}))}bindUnits(){this.dataUnitComponents.map((t=>{t.units.map((t=>{t.addEventListener("click",(()=>{let e=t.getAttribute("data-point-index");this.setCurrentDataPoint(e)}))}))})),this.tip.container.addEventListener("click",(()=>{let t=this.tip.container.getAttribute("data-point-index");this.setCurrentDataPoint(t)}))}updateOverlay(){this.overlayGuides.map((t=>{let e=t.units[this.state.currentIndex];W[t.type](e,t.overlay)}))}onLeftArrow(){this.setCurrentDataPoint(this.state.currentIndex-1)}onRightArrow(){this.setCurrentDataPoint(this.state.currentIndex+1)}getDataPoint(t=this.state.currentIndex){let e=this.state;return{index:t,label:e.xAxis.labels[t],values:e.datasets.map((e=>e.values[t]))}}setCurrentDataPoint(t){let e=this.state;(t=parseInt(t))<0&&(t=0),t>=e.xAxis.labels.length&&(t=e.xAxis.labels.length-1),t!==e.currentIndex&&(e.currentIndex=t,function(t,e,s){var i=document.createEvent("HTMLEvents");for(var a in i.initEvent(e,!0,!0),s)i[a]=s[a];t.dispatchEvent(i)}(this.parent,"data-select",this.getDataPoint()))}addDataPoint(t,e,s=this.state.datasetLength){super.addDataPoint(t,e,s),this.data.labels.splice(s,0,t),this.data.datasets.map(((t,i)=>{t.values.splice(s,0,e[i])})),this.update(this.data)}removeDataPoint(t=this.state.datasetLength-1){this.data.labels.length<=1||(super.removeDataPoint(t),this.data.labels.splice(t,1),this.data.datasets.map((e=>{e.values.splice(t,1)})),this.update(this.data))}updateDataset(t,e=0){this.data.datasets[e].values=t,this.update(this.data)}updateDatasets(t){this.data.datasets.map(((e,s)=>{t[s]&&(e.values=t[s])})),this.update(this.data)}}const dt={bar:ht,line:ht,percentage:class extends V{constructor(t,e){super(t,e),this.type="percentage",this.setup()}setMeasures(t){let e=this.measures;this.barOptions=t.barOptions||{};let s=this.barOptions;s.height=s.height||20,s.depth=s.depth||2,e.paddings.right=30,e.legendHeight=60,e.baseHeight=8*(s.height+.5*s.depth)}setupComponents(){let t=this.state,e=[["percentageBars",{barHeight:this.barOptions.height,barDepth:this.barOptions.depth},function(){return{xPositions:t.xPositions,widths:t.widths,colors:this.colors}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calc(){super.calc();let t=this.state;t.xPositions=[],t.widths=[];let e=0;t.sliceTotals.map((s=>{let i=this.width*s/t.grandTotal;t.widths.push(i),t.xPositions.push(e),e+=i}))}makeDataByIndex(){}bindTooltip(){let t=this.state;this.container.addEventListener("mousemove",(s=>{let i=this.components.get("percentageBars").store,a=s.target;if(i.includes(a)){let s=i.indexOf(a),n=e(this.container),r=e(a),o=r.left-n.left+parseInt(a.getAttribute("width"))/2,l=r.top-n.top,h=(this.formattedLabels&&this.formattedLabels.length>0?this.formattedLabels[s]:this.state.labels[s])+": ",d=t.sliceTotals[s]/t.grandTotal;this.tip.setValues(o,l,{name:h,value:(100*d).toFixed(1)+"%"}),this.tip.showTip()}}))}},heatmap:class extends B{constructor(t,e){super(t,e),this.type="heatmap",this.countLabel=e.countLabel||"";let s=["Sunday","Monday"],i=s.includes(e.startSubDomain)?e.startSubDomain:"Sunday";this.startSubDomainIndex=s.indexOf(i),this.setup()}setMeasures(t){let e=this.measures;this.discreteDomains=0===t.discreteDomains?0:1,e.paddings.top=36,e.paddings.bottom=0,e.legendHeight=24,e.baseHeight=84+n(e);let s=this.data,i=this.discreteDomains?12:0;this.independentWidth=12*(J(s.start,s.end)+i)+r(e)}updateWidth(){let t=this.discreteDomains?12:0,e=this.state.noOfWeeks?this.state.noOfWeeks:52;this.baseWidth=12*(e+t)+r(this.measures)}prepareData(t=this.data){if(t.start&&t.end&&t.start>t.end)throw new Error("Start date cannot be greater than end date.");if(t.start||(t.start=new Date,t.start.setFullYear(t.start.getFullYear()-1)),t.end||(t.end=new Date),t.dataPoints=t.dataPoints||{},parseInt(Object.keys(t.dataPoints)[0])>1e5){let e={};Object.keys(t.dataPoints).forEach((s=>{let i=new Date(1e3*s);e[X(i)]=t.dataPoints[s]})),t.dataPoints=e}return t}calc(){let t=this.state;t.start=_(this.data.start),t.end=_(this.data.end),t.firstWeekStart=_(t.start),t.noOfWeeks=J(t.start,t.end),t.distribution=function(t,e){let s=Math.max(...t),i=1/(e-1),a=[];for(var n=0;n<e;n++){let t=s*(i*n);a.push(t)}return a}(Object.values(this.data.dataPoints),5),t.domainConfigs=this.getDomains()}setupComponents(){let t=this.state,e=this.discreteDomains?0:1,s=t.domainConfigs.map(((s,i)=>["heatDomain",{index:s.index,colWidth:12,rowHeight:12,squareSize:10,radius:this.rawChartArgs.radius||0,xTranslate:12*t.domainConfigs.filter(((t,e)=>e<i)).map((t=>t.cols.length-e)).reduce(((t,e)=>t+e),0)},function(){return t.domainConfigs[i]}.bind(this)]));this.components=new Map(s.map(((t,e)=>{let s=at(...t);return[t[0]+"-"+e,s]})));let i=0;G.forEach(((t,e)=>{if([1,3,5].includes(e)){let e=E("subdomain-name",-6,i,t,{fontSize:10,dy:8,textAnchor:"end"});this.drawArea.appendChild(e)}i+=12}))}update(t){t||console.error("No data to update."),this.data=this.prepareData(t),this.draw(),this.bindTooltip()}bindTooltip(){this.container.addEventListener("mousemove",(t=>{this.components.forEach((e=>{let s=e.store,i=t.target;if(s.includes(i)){let e=i.getAttribute("data-value"),s=i.getAttribute("data-date").split("-"),a=Q(parseInt(s[1])-1,!0),n=this.container.getBoundingClientRect(),r=i.getBoundingClientRect(),o=parseInt(t.target.getAttribute("width")),l=r.left-n.left+o/2,h=r.top-n.top,d=e+" "+this.countLabel,c=" on "+a+" "+s[0]+", "+s[2];this.tip.setValues(l,h,{name:c,value:d,valueFirst:1},[]),this.tip.showTip()}}))}))}renderLegend(){this.legendArea.textContent="";let t=0,e=this.rawChartArgs.radius||0,s=E("subdomain-name",t,12,"Less",{fontSize:11,dy:9});t=30,this.legendArea.appendChild(s),this.colors.slice(0,5).map(((s,i)=>{const a=N("heatmap-legend-unit",t+15*i,12,10,e,s);this.legendArea.appendChild(a)}));let i=E("subdomain-name",t+75+3,12,"More",{fontSize:11,dy:9});this.legendArea.appendChild(i)}getDomains(){let t=this.state;const[e,s]=[t.start.getMonth(),t.start.getFullYear()],[i,a]=[t.end.getMonth(),t.end.getFullYear()],n=i-e+1+12*(a-s);let r=[],o=_(t.start);for(var l=0;l<n;l++){let e=t.end;if(!K(o,t.end)){let[t,s]=[o.getMonth(),o.getFullYear()];e=Z(t,s)}r.push(this.getDomainConfig(o,e)),et(e,1),o=e}return r}getDomainConfig(t,e=""){let[s,i]=[t.getMonth(),t.getFullYear()],a=tt(t),n={index:s,cols:[]};et(e=_(e)||Z(s,i),1);let r,o=J(a,e),l=[];for(var h=0;h<o;h++)r=this.getCol(a,s),l.push(r),a=new Date(r[6].yyyyMmDd),et(a,1);return void 0!==r[6].dataValue&&(et(a,1),l.push(this.getCol(a,s,!0))),n.cols=l,n}getCol(t,e,s=!1){let i=this.state,a=_(t),n=[];for(var r=0;r<7;r++,et(a,1)){let t={},r=a>=i.start&&a<=i.end;s||a.getMonth()!==e||!r?t.yyyyMmDd=X(a):t=this.getSubDomainConfig(a),n.push(t)}return n}getSubDomainConfig(t){let e=X(t),s=this.data.dataPoints[e];var i,a;return{yyyyMmDd:e,dataValue:s||0,fill:this.colors[(i=s,a=this.state.distribution,a.filter((t=>t<i)).length)]}}},pie:class extends V{constructor(t,e){super(t,e),this.type="pie",this.initTimeout=0,this.init=1,this.setup()}configure(t){super.configure(t),this.mouseMove=this.mouseMove.bind(this),this.mouseLeave=this.mouseLeave.bind(this),this.hoverRadio=t.hoverRadio||.1,this.config.startAngle=t.startAngle||0,this.clockWise=t.clockWise||!1}calc(){super.calc();let t=this.state;this.radius=this.height>this.width?this.center.x:this.center.y;const{radius:e,clockWise:s}=this,i=t.slicesProperties||[];t.sliceStrings=[],t.slicesProperties=[];let a=180-this.config.startAngle;t.sliceTotals.map(((n,r)=>{const o=a,l=n/t.grandTotal*360,h=l>180?1:0,d=s?-l:l,c=a+=d,p=m(o,e),u=m(c,e),g=this.init&&i[r];let f,y;this.init?(f=g?g.startPosition:p,y=g?g.endPosition:p):(f=p,y=u);const b=360===l?function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h,d]=[s.x+e.x,2*s.y,s.y+e.y];return`M${s.x} ${s.y}\n\t\tL${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h} z\n\t\tL${r} ${h}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${d} z`}(f,y,this.center,this.radius,s,h):function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h]=[s.x+e.x,s.y+e.y];return`M${s.x} ${s.y}\n\t\tL${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h} z`}(f,y,this.center,this.radius,s,h);t.sliceStrings.push(b),t.slicesProperties.push({startPosition:p,endPosition:u,value:n,total:t.grandTotal,startAngle:o,endAngle:c,angle:d})})),this.init=0}setupComponents(){let t=this.state,e=[["pieSlices",{},function(){return{sliceStrings:t.sliceStrings,colors:this.colors}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calTranslateByAngle(t){const{radius:e,hoverRadio:s}=this,i=m(t.startAngle+t.angle/2,e);return`translate3d(${i.x*s}px,${i.y*s}px,0)`}hoverSlice(t,s,i,a){if(!t)return;const n=this.colors[s];if(i){Y(t,this.calTranslateByAngle(this.state.slicesProperties[s])),t.style.fill=k(n,50);let i=e(this.svg),r=a.pageX-i.left+10,o=a.pageY-i.top-10,l=(this.formatted_labels&&this.formatted_labels.length>0?this.formatted_labels[s]:this.state.labels[s])+": ",h=(100*this.state.sliceTotals[s]/this.state.grandTotal).toFixed(1);this.tip.setValues(r,o,{name:l,value:h+"%"}),this.tip.showTip()}else Y(t,"translate3d(0,0,0)"),this.tip.hideTip(),t.style.fill=n}bindTooltip(){this.container.addEventListener("mousemove",this.mouseMove),this.container.addEventListener("mouseleave",this.mouseLeave)}mouseMove(t){const e=t.target;let s=this.components.get("pieSlices").store,i=this.curActiveSliceIndex,a=this.curActiveSlice;if(s.includes(e)){let n=s.indexOf(e);this.hoverSlice(a,i,!1),this.curActiveSlice=e,this.curActiveSliceIndex=n,this.hoverSlice(e,n,!0,t)}else this.mouseLeave()}mouseLeave(){this.hoverSlice(this.curActiveSlice,this.curActiveSliceIndex,!1)}},donut:class extends V{constructor(t,e){super(t,e),this.type="donut",this.initTimeout=0,this.init=1,this.setup()}configure(t){super.configure(t),this.mouseMove=this.mouseMove.bind(this),this.mouseLeave=this.mouseLeave.bind(this),this.hoverRadio=t.hoverRadio||.1,this.config.startAngle=t.startAngle||0,this.clockWise=t.clockWise||!1,this.strokeWidth=t.strokeWidth||30}calc(){super.calc();let t=this.state;this.radius=this.height>this.width?this.center.x-this.strokeWidth/2:this.center.y-this.strokeWidth/2;const{radius:e,clockWise:s}=this,i=t.slicesProperties||[];t.sliceStrings=[],t.slicesProperties=[];let a=180-this.config.startAngle;t.sliceTotals.map(((n,r)=>{const o=a,l=n/t.grandTotal*360,h=l>180?1:0,d=s?-l:l,c=a+=d,p=m(o,e),u=m(c,e),g=this.init&&i[r];let f,y;this.init?(f=g?g.startPosition:p,y=g?g.endPosition:p):(f=p,y=u);const b=360===l?function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h,d]=[s.x+e.x,2*i+o,s.y+t.y];return`M${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h}\n\t\tM${r} ${h}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${d}`}(f,y,this.center,this.radius,this.clockWise,h):function(t,e,s,i,a=1,n=0){let[r,o]=[s.x+t.x,s.y+t.y],[l,h]=[s.x+e.x,s.y+e.y];return`M${r} ${o}\n\t\tA ${i} ${i} 0 ${n} ${a?1:0}\n\t\t${l} ${h}`}(f,y,this.center,this.radius,this.clockWise,h);t.sliceStrings.push(b),t.slicesProperties.push({startPosition:p,endPosition:u,value:n,total:t.grandTotal,startAngle:o,endAngle:c,angle:d})})),this.init=0}setupComponents(){let t=this.state,e=[["donutSlices",{},function(){return{sliceStrings:t.sliceStrings,colors:this.colors,strokeWidth:this.strokeWidth}}.bind(this)]];this.components=new Map(e.map((t=>{let e=at(...t);return[t[0],e]})))}calTranslateByAngle(t){const{radius:e,hoverRadio:s}=this,i=m(t.startAngle+t.angle/2,e);return`translate3d(${i.x*s}px,${i.y*s}px,0)`}hoverSlice(t,s,i,a){if(!t)return;const n=this.colors[s];if(i){Y(t,this.calTranslateByAngle(this.state.slicesProperties[s])),t.style.stroke=k(n,50);let i=e(this.svg),r=a.pageX-i.left+10,o=a.pageY-i.top-10,l=(this.formatted_labels&&this.formatted_labels.length>0?this.formatted_labels[s]:this.state.labels[s])+": ",h=(100*this.state.sliceTotals[s]/this.state.grandTotal).toFixed(1);this.tip.setValues(r,o,{name:l,value:h+"%"}),this.tip.showTip()}else Y(t,"translate3d(0,0,0)"),this.tip.hideTip(),t.style.stroke=n}bindTooltip(){this.container.addEventListener("mousemove",this.mouseMove),this.container.addEventListener("mouseleave",this.mouseLeave)}mouseMove(t){const e=t.target;let s=this.components.get("donutSlices").store,i=this.curActiveSliceIndex,a=this.curActiveSlice;if(s.includes(e)){let n=s.indexOf(e);this.hoverSlice(a,i,!1),this.curActiveSlice=e,this.curActiveSliceIndex=n,this.hoverSlice(e,n,!0,t)}else this.mouseLeave()}mouseLeave(){this.hoverSlice(this.curActiveSlice,this.curActiveSliceIndex,!1)}}};class ct{constructor(t,e){return function(t="line",e,s){return"axis-mixed"===t?(s.type="line",new ht(e,s)):dt[t]?new dt[t](e,s):void console.error("Undefined chart type: "+t)}(e.type,t,e)}}const pt=document.getElementById("main"),ut=JSON.parse(document.querySelector(".data-json").innerHTML);let mt;const gt=(t,e)=>{const s={Fast:[],Average:[],Slow:[]},i=[];return Object.keys(ut).forEach((i=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((a=>{ut[i][a]&&ut[i][a].forEach((i=>{i&&i.client&&i.client===e&&i.platform&&i.platform.toLowerCase()===t.toLowerCase()&&(s.Slow.push(i.slow),s.Average.push(i.avg),s.Fast.push(i.fast))}))}))})),["Slow","Average","Fast"].forEach((t=>{i.push({name:t,chartType:"bar",values:s[t]})})),i},ft=()=>{const t=[];return Object.keys(ut).forEach((e=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((s=>{ut[e][s]&&t.push(`${s}/'${e.substr(2,2)}`)}))})),t},yt={yRegions:[{label:"%",start:0,end:100}],data:{labels:ft()},barOptions:{spaceRatio:.001,stacked:!0},height:400,colors:["red","yellow","green","black"],axisOptions:{xAxisMode:"tick",yAxisMode:"span"}},bt=(t,e)=>{mt?mt.update({labels:ft(),datasets:gt(t,e)}):(yt.data.datasets=gt(t,e),mt=new ct(pt,yt))},xt=(()=>{const t=[];return Object.keys(ut).forEach((e=>{[1,2,3,4,5,6,7,8,9,10,11,12].forEach((s=>{ut[e][s]&&ut[e][s].forEach((e=>{e.platform&&!t.includes(e.platform.toLowerCase())&&t.push(e.platform.toLowerCase())}))}))})),t})(),vt=document.createElement("select"),wt=document.createElement("select");wt.innerHTML='<option value="mobile">Mobile</option><option value="desktop">Desktop</option>',xt.sort().forEach((t=>{const e=document.createElement("option");var s;e.text=(s=t)&&s[0].toUpperCase()+s.slice(1),e.value=t,vt.appendChild(e)})),vt.addEventListener("change",(t=>{bt(t.target.value,wt.value)})),wt.addEventListener("change",(t=>{bt(vt.value,t.target.value)})),pt.insertAdjacentElement("beforebegin",vt),pt.insertAdjacentElement("beforebegin",wt),bt(xt[0],"mobile");
